added UI options for Webcams
Don't know how to activate it tho
