--- GUI/scAnt_GUI_mw.py
@@ -1,987 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'I:\3D_Scanner\scAnt\GUI\scAnt_GUI_mw_grid.ui'
-#
-# Created by: PyQt5 UI code generator 5.13.0
-#
-# WARNING! All changes made in this file will be lost!
-
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-
-class Ui_MainWindow(object):
-    def setupUi(self, MainWindow):
-        MainWindow.setObjectName("MainWindow")
-        MainWindow.setWindowModality(QtCore.Qt.NonModal)
-        MainWindow.setEnabled(True)
-        MainWindow.resize(1600, 936)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(MainWindow.sizePolicy().hasHeightForWidth())
-        MainWindow.setSizePolicy(sizePolicy)
-        MainWindow.setMinimumSize(QtCore.QSize(1600, 936))
-        MainWindow.setMaximumSize(QtCore.QSize(3840, 2160))
-        MainWindow.setWindowTitle("scAnt")
-        self.centralwidget = QtWidgets.QWidget(MainWindow)
-        self.centralwidget.setMinimumSize(QtCore.QSize(1600, 900))
-        self.centralwidget.setObjectName("centralwidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.centralwidget)
-        self.gridLayout.setObjectName("gridLayout")
-        self.frame_info = QtWidgets.QFrame(self.centralwidget)
-        self.frame_info.setMinimumSize(QtCore.QSize(370, 0))
-        self.frame_info.setFrameShape(QtWidgets.QFrame.Box)
-        self.frame_info.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_info.setObjectName("frame_info")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.frame_info)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.label_info = QtWidgets.QLabel(self.frame_info)
-        self.label_info.setMaximumSize(QtCore.QSize(16777215, 30))
-        font = QtGui.QFont()
-        font.setPointSize(11)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_info.setFont(font)
-        self.label_info.setObjectName("label_info")
-        self.verticalLayout.addWidget(self.label_info)
-        self.label_progressTotal = QtWidgets.QLabel(self.frame_info)
-        self.label_progressTotal.setEnabled(False)
-        self.label_progressTotal.setMaximumSize(QtCore.QSize(16777215, 30))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_progressTotal.setFont(font)
-        self.label_progressTotal.setObjectName("label_progressTotal")
-        self.verticalLayout.addWidget(self.label_progressTotal)
-        self.progressBar_total = QtWidgets.QProgressBar(self.frame_info)
-        self.progressBar_total.setEnabled(False)
-        self.progressBar_total.setMaximumSize(QtCore.QSize(16777215, 40))
-        self.progressBar_total.setProperty("value", 0)
-        self.progressBar_total.setObjectName("progressBar_total")
-        self.verticalLayout.addWidget(self.progressBar_total)
-        self.label_log = QtWidgets.QLabel(self.frame_info)
-        self.label_log.setMaximumSize(QtCore.QSize(16777215, 30))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_log.setFont(font)
-        self.label_log.setObjectName("label_log")
-        self.verticalLayout.addWidget(self.label_log)
-        self.listWidget_log = QtWidgets.QListWidget(self.frame_info)
-        self.listWidget_log.setMinimumSize(QtCore.QSize(0, 200))
-        self.listWidget_log.setMaximumSize(QtCore.QSize(1000, 400))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        self.listWidget_log.setFont(font)
-        self.listWidget_log.setFlow(QtWidgets.QListView.TopToBottom)
-        self.listWidget_log.setObjectName("listWidget_log")
-        self.verticalLayout.addWidget(self.listWidget_log)
-        self.gridLayout.addWidget(self.frame_info, 2, 2, 1, 1)
-        self.frame_cameraSettings = QtWidgets.QFrame(self.centralwidget)
-        self.frame_cameraSettings.setMinimumSize(QtCore.QSize(370, 0))
-        self.frame_cameraSettings.setFrameShape(QtWidgets.QFrame.Box)
-        self.frame_cameraSettings.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_cameraSettings.setObjectName("frame_cameraSettings")
-        self.gridLayout_3 = QtWidgets.QGridLayout(self.frame_cameraSettings)
-        self.gridLayout_3.setObjectName("gridLayout_3")
-        self.pushButton_captureImage = QtWidgets.QPushButton(self.frame_cameraSettings)
-        self.pushButton_captureImage.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.pushButton_captureImage.setFont(font)
-        self.pushButton_captureImage.setObjectName("pushButton_captureImage")
-        self.gridLayout_3.addWidget(self.pushButton_captureImage, 4, 1, 1, 1)
-        self.comboBox_selectCamera = QtWidgets.QComboBox(self.frame_cameraSettings)
-        self.comboBox_selectCamera.setObjectName("comboBox_selectCamera")
-        self.gridLayout_3.addWidget(self.comboBox_selectCamera, 2, 1, 1, 1)
-        self.label_cameraSettings = QtWidgets.QLabel(self.frame_cameraSettings)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Minimum)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_cameraSettings.sizePolicy().hasHeightForWidth())
-        self.label_cameraSettings.setSizePolicy(sizePolicy)
-        self.label_cameraSettings.setMinimumSize(QtCore.QSize(0, 40))
-        self.label_cameraSettings.setMaximumSize(QtCore.QSize(16777215, 50))
-        font = QtGui.QFont()
-        font.setPointSize(11)
-        font.setBold(True)
-        font.setWeight(75)
-        font.setKerning(True)
-        self.label_cameraSettings.setFont(font)
-        self.label_cameraSettings.setObjectName("label_cameraSettings")
-        self.gridLayout_3.addWidget(self.label_cameraSettings, 0, 0, 1, 2)
-        self.pushButton_startLiveView = QtWidgets.QPushButton(self.frame_cameraSettings)
-        self.pushButton_startLiveView.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.pushButton_startLiveView.setFont(font)
-        self.pushButton_startLiveView.setObjectName("pushButton_startLiveView")
-        self.gridLayout_3.addWidget(self.pushButton_startLiveView, 4, 0, 1, 1)
-        self.label_selectCamera = QtWidgets.QLabel(self.frame_cameraSettings)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_selectCamera.setFont(font)
-        self.label_selectCamera.setObjectName("label_selectCamera")
-        self.gridLayout_3.addWidget(self.label_selectCamera, 2, 0, 1, 1)
-        self.stacked_camera_settings = QtWidgets.QStackedWidget(self.frame_cameraSettings)
-        self.stacked_camera_settings.setAccessibleDescription("")
-        self.stacked_camera_settings.setObjectName("stacked_camera_settings")
-        self.FLIR = QtWidgets.QWidget()
-        self.FLIR.setObjectName("FLIR")
-        self.gridLayout_5 = QtWidgets.QGridLayout(self.FLIR)
-        self.gridLayout_5.setObjectName("gridLayout_5")
-        self.label_exposureAuto = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_exposureAuto.setFont(font)
-        self.label_exposureAuto.setObjectName("label_exposureAuto")
-        self.gridLayout_5.addWidget(self.label_exposureAuto, 0, 0, 1, 1)
-        self.checkBox_exposureAuto = QtWidgets.QCheckBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.checkBox_exposureAuto.setFont(font)
-        self.checkBox_exposureAuto.setText("")
-        self.checkBox_exposureAuto.setChecked(False)
-        self.checkBox_exposureAuto.setObjectName("checkBox_exposureAuto")
-        self.gridLayout_5.addWidget(self.checkBox_exposureAuto, 0, 1, 1, 1)
-        self.label_exposureTime = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_exposureTime.setFont(font)
-        self.label_exposureTime.setObjectName("label_exposureTime")
-        self.gridLayout_5.addWidget(self.label_exposureTime, 1, 0, 1, 1)
-        self.doubleSpinBox_exposureTime = QtWidgets.QDoubleSpinBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_exposureTime.setFont(font)
-        self.doubleSpinBox_exposureTime.setDecimals(0)
-        self.doubleSpinBox_exposureTime.setMinimum(100.0)
-        self.doubleSpinBox_exposureTime.setMaximum(1000000.0)
-        self.doubleSpinBox_exposureTime.setSingleStep(1000.0)
-        self.doubleSpinBox_exposureTime.setProperty("value", 90000.0)
-        self.doubleSpinBox_exposureTime.setObjectName("doubleSpinBox_exposureTime")
-        self.gridLayout_5.addWidget(self.doubleSpinBox_exposureTime, 1, 1, 1, 1)
-        self.label_gainAuto = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_gainAuto.setFont(font)
-        self.label_gainAuto.setObjectName("label_gainAuto")
-        self.gridLayout_5.addWidget(self.label_gainAuto, 2, 0, 1, 1)
-        self.checkBox_gainAuto = QtWidgets.QCheckBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.checkBox_gainAuto.setFont(font)
-        self.checkBox_gainAuto.setText("")
-        self.checkBox_gainAuto.setObjectName("checkBox_gainAuto")
-        self.gridLayout_5.addWidget(self.checkBox_gainAuto, 2, 1, 1, 1)
-        self.label_gainLevel = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_gainLevel.setFont(font)
-        self.label_gainLevel.setObjectName("label_gainLevel")
-        self.gridLayout_5.addWidget(self.label_gainLevel, 3, 0, 1, 1)
-        self.doubleSpinBox_gainLevel = QtWidgets.QDoubleSpinBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_gainLevel.setFont(font)
-        self.doubleSpinBox_gainLevel.setDecimals(2)
-        self.doubleSpinBox_gainLevel.setMinimum(0.0)
-        self.doubleSpinBox_gainLevel.setMaximum(25.0)
-        self.doubleSpinBox_gainLevel.setProperty("value", 1.83)
-        self.doubleSpinBox_gainLevel.setObjectName("doubleSpinBox_gainLevel")
-        self.gridLayout_5.addWidget(self.doubleSpinBox_gainLevel, 3, 1, 1, 1)
-        self.label_gamma = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_gamma.setFont(font)
-        self.label_gamma.setObjectName("label_gamma")
-        self.gridLayout_5.addWidget(self.label_gamma, 4, 0, 1, 1)
-        self.doubleSpinBox_gamma = QtWidgets.QDoubleSpinBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_gamma.setFont(font)
-        self.doubleSpinBox_gamma.setDecimals(2)
-        self.doubleSpinBox_gamma.setMinimum(0.1)
-        self.doubleSpinBox_gamma.setMaximum(4.0)
-        self.doubleSpinBox_gamma.setSingleStep(0.1)
-        self.doubleSpinBox_gamma.setProperty("value", 0.8)
-        self.doubleSpinBox_gamma.setObjectName("doubleSpinBox_gamma")
-        self.gridLayout_5.addWidget(self.doubleSpinBox_gamma, 4, 1, 1, 1)
-        self.label_balanceRatioRed = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_balanceRatioRed.setFont(font)
-        self.label_balanceRatioRed.setObjectName("label_balanceRatioRed")
-        self.gridLayout_5.addWidget(self.label_balanceRatioRed, 5, 0, 1, 1)
-        self.doubleSpinBox_balanceRatioRed = QtWidgets.QDoubleSpinBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_balanceRatioRed.setFont(font)
-        self.doubleSpinBox_balanceRatioRed.setDecimals(2)
-        self.doubleSpinBox_balanceRatioRed.setMinimum(0.25)
-        self.doubleSpinBox_balanceRatioRed.setMaximum(4.0)
-        self.doubleSpinBox_balanceRatioRed.setSingleStep(0.1)
-        self.doubleSpinBox_balanceRatioRed.setProperty("value", 1.58)
-        self.doubleSpinBox_balanceRatioRed.setObjectName("doubleSpinBox_balanceRatioRed")
-        self.gridLayout_5.addWidget(self.doubleSpinBox_balanceRatioRed, 5, 1, 1, 1)
-        self.label_balanceRatioBlue = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_balanceRatioBlue.setFont(font)
-        self.label_balanceRatioBlue.setObjectName("label_balanceRatioBlue")
-        self.gridLayout_5.addWidget(self.label_balanceRatioBlue, 6, 0, 1, 1)
-        self.doubleSpinBox_balanceRatioBlue = QtWidgets.QDoubleSpinBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_balanceRatioBlue.setFont(font)
-        self.doubleSpinBox_balanceRatioBlue.setDecimals(2)
-        self.doubleSpinBox_balanceRatioBlue.setMinimum(0.25)
-        self.doubleSpinBox_balanceRatioBlue.setMaximum(4.0)
-        self.doubleSpinBox_balanceRatioBlue.setSingleStep(0.1)
-        self.doubleSpinBox_balanceRatioBlue.setProperty("value", 1.79)
-        self.doubleSpinBox_balanceRatioBlue.setObjectName("doubleSpinBox_balanceRatioBlue")
-        self.gridLayout_5.addWidget(self.doubleSpinBox_balanceRatioBlue, 6, 1, 1, 1)
-        self.label_highlightExposure = QtWidgets.QLabel(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_highlightExposure.setFont(font)
-        self.label_highlightExposure.setObjectName("label_highlightExposure")
-        self.gridLayout_5.addWidget(self.label_highlightExposure, 7, 0, 1, 1)
-        self.checkBox_highlightExposure = QtWidgets.QCheckBox(self.FLIR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.checkBox_highlightExposure.setFont(font)
-        self.checkBox_highlightExposure.setText("")
-        self.checkBox_highlightExposure.setObjectName("checkBox_highlightExposure")
-        self.gridLayout_5.addWidget(self.checkBox_highlightExposure, 7, 1, 1, 1)
-        self.stacked_camera_settings.addWidget(self.FLIR)
-        self.DSLR = QtWidgets.QWidget()
-        self.DSLR.setObjectName("DSLR")
-        self.gridLayout_7 = QtWidgets.QGridLayout(self.DSLR)
-        self.gridLayout_7.setObjectName("gridLayout_7")
-        self.label_shutterSpeed = QtWidgets.QLabel(self.DSLR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_shutterSpeed.setFont(font)
-        self.label_shutterSpeed.setObjectName("label_shutterSpeed")
-        self.gridLayout_7.addWidget(self.label_shutterSpeed, 0, 0, 1, 1)
-        self.comboBox_shutterSpeed = QtWidgets.QComboBox(self.DSLR)
-        self.comboBox_shutterSpeed.setObjectName("comboBox_shutterSpeed")
-        self.gridLayout_7.addWidget(self.comboBox_shutterSpeed, 0, 1, 1, 1)
-        self.label_aperture = QtWidgets.QLabel(self.DSLR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_aperture.setFont(font)
-        self.label_aperture.setObjectName("label_aperture")
-        self.gridLayout_7.addWidget(self.label_aperture, 1, 0, 1, 1)
-        self.comboBox_aperture = QtWidgets.QComboBox(self.DSLR)
-        self.comboBox_aperture.setObjectName("comboBox_aperture")
-        self.gridLayout_7.addWidget(self.comboBox_aperture, 1, 1, 1, 1)
-        self.label_iso = QtWidgets.QLabel(self.DSLR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_iso.setFont(font)
-        self.label_iso.setObjectName("label_iso")
-        self.gridLayout_7.addWidget(self.label_iso, 2, 0, 1, 1)
-        self.comboBox_iso = QtWidgets.QComboBox(self.DSLR)
-        self.comboBox_iso.setObjectName("comboBox_iso")
-        self.gridLayout_7.addWidget(self.comboBox_iso, 2, 1, 1, 1)
-        self.label_whiteBalance = QtWidgets.QLabel(self.DSLR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_whiteBalance.setFont(font)
-        self.label_whiteBalance.setObjectName("label_whiteBalance")
-        self.gridLayout_7.addWidget(self.label_whiteBalance, 3, 0, 1, 1)
-        self.comboBox_whiteBalance = QtWidgets.QComboBox(self.DSLR)
-        self.comboBox_whiteBalance.setObjectName("comboBox_whiteBalance")
-        self.gridLayout_7.addWidget(self.comboBox_whiteBalance, 3, 1, 1, 1)
-        self.label_compression = QtWidgets.QLabel(self.DSLR)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_compression.setFont(font)
-        self.label_compression.setObjectName("label_compression")
-        self.gridLayout_7.addWidget(self.label_compression, 4, 0, 1, 1)
-        self.comboBox_compression = QtWidgets.QComboBox(self.DSLR)
-        self.comboBox_compression.setObjectName("comboBox_compression")
-        self.gridLayout_7.addWidget(self.comboBox_compression, 4, 1, 1, 1)
-        self.stacked_camera_settings.addWidget(self.DSLR)
-        self.gridLayout_3.addWidget(self.stacked_camera_settings, 3, 0, 1, 2)
-        self.gridLayout.addWidget(self.frame_cameraSettings, 1, 1, 1, 1)
-        self.frame_stepperControl = QtWidgets.QFrame(self.centralwidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.frame_stepperControl.sizePolicy().hasHeightForWidth())
-        self.frame_stepperControl.setSizePolicy(sizePolicy)
-        self.frame_stepperControl.setMinimumSize(QtCore.QSize(0, 350))
-        self.frame_stepperControl.setFrameShape(QtWidgets.QFrame.Box)
-        self.frame_stepperControl.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_stepperControl.setObjectName("frame_stepperControl")
-        self.gridLayout_2 = QtWidgets.QGridLayout(self.frame_stepperControl)
-        self.gridLayout_2.setHorizontalSpacing(10)
-        self.gridLayout_2.setObjectName("gridLayout_2")
-        self.label_yAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_yAxisSliderMax.setFont(font)
-        self.label_yAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.label_yAxisSliderMax.setObjectName("label_yAxisSliderMax")
-        self.gridLayout_2.addWidget(self.label_yAxisSliderMax, 7, 1, 1, 1)
-        self.label_setRangeXMax = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeXMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeXMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeXMax.setFont(font)
-        self.label_setRangeXMax.setObjectName("label_setRangeXMax")
-        self.gridLayout_2.addWidget(self.label_setRangeXMax, 4, 6, 1, 1)
-        self.horizontalSlider_xAxis = QtWidgets.QSlider(self.frame_stepperControl)
-        self.horizontalSlider_xAxis.setMinimumSize(QtCore.QSize(500, 0))
-        self.horizontalSlider_xAxis.setMaximum(450)
-        self.horizontalSlider_xAxis.setOrientation(QtCore.Qt.Horizontal)
-        self.horizontalSlider_xAxis.setObjectName("horizontalSlider_xAxis")
-        self.gridLayout_2.addWidget(self.horizontalSlider_xAxis, 5, 0, 1, 2)
-        self.doubleSpinBox_xMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_xMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_xMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_xMin.setFont(font)
-        self.doubleSpinBox_xMin.setDecimals(0)
-        self.doubleSpinBox_xMin.setMinimum(0.0)
-        self.doubleSpinBox_xMin.setMaximum(450.0)
-        self.doubleSpinBox_xMin.setSingleStep(5.0)
-        self.doubleSpinBox_xMin.setProperty("value", 0.0)
-        self.doubleSpinBox_xMin.setObjectName("doubleSpinBox_xMin")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_xMin, 5, 4, 1, 1)
-        self.label_setRangeZMax = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeZMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeZMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeZMax.setFont(font)
-        self.label_setRangeZMax.setObjectName("label_setRangeZMax")
-        self.gridLayout_2.addWidget(self.label_setRangeZMax, 10, 6, 1, 1)
-        self.label_setRangeYMin = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeYMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeYMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeYMin.setFont(font)
-        self.label_setRangeYMin.setObjectName("label_setRangeYMin")
-        self.gridLayout_2.addWidget(self.label_setRangeYMin, 7, 4, 1, 1)
-        self.doubleSpinBox_zMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_zMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_zMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_zMin.setFont(font)
-        self.doubleSpinBox_zMin.setDecimals(0)
-        self.doubleSpinBox_zMin.setMinimum(-45000.0)
-        self.doubleSpinBox_zMin.setMaximum(0.0)
-        self.doubleSpinBox_zMin.setSingleStep(1000.0)
-        self.doubleSpinBox_zMin.setProperty("value", -20000.0)
-        self.doubleSpinBox_zMin.setObjectName("doubleSpinBox_zMin")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_zMin, 11, 4, 1, 1)
-        self.label_setRangeZMin = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeZMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeZMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeZMin.setFont(font)
-        self.label_setRangeZMin.setObjectName("label_setRangeZMin")
-        self.gridLayout_2.addWidget(self.label_setRangeZMin, 10, 4, 1, 1)
-        self.label_setRangeYMax = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeYMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeYMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeYMax.setFont(font)
-        self.label_setRangeYMax.setObjectName("label_setRangeYMax")
-        self.gridLayout_2.addWidget(self.label_setRangeYMax, 7, 6, 1, 1)
-        self.label_xAxis = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_xAxis.setFont(font)
-        self.label_xAxis.setObjectName("label_xAxis")
-        self.gridLayout_2.addWidget(self.label_xAxis, 3, 0, 1, 1)
-        self.label_setRangeXStep = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeXStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeXStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeXStep.setFont(font)
-        self.label_setRangeXStep.setObjectName("label_setRangeXStep")
-        self.gridLayout_2.addWidget(self.label_setRangeXStep, 4, 5, 1, 1)
-        self.horizontalSlider_zAxis = QtWidgets.QSlider(self.frame_stepperControl)
-        self.horizontalSlider_zAxis.setMinimumSize(QtCore.QSize(500, 0))
-        self.horizontalSlider_zAxis.setMinimum(-45000)
-        self.horizontalSlider_zAxis.setMaximum(0)
-        self.horizontalSlider_zAxis.setOrientation(QtCore.Qt.Horizontal)
-        self.horizontalSlider_zAxis.setObjectName("horizontalSlider_zAxis")
-        self.gridLayout_2.addWidget(self.horizontalSlider_zAxis, 11, 0, 1, 2)
-        self.lcdNumber_yAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
-        self.lcdNumber_yAxis.setMinimumSize(QtCore.QSize(70, 0))
-        self.lcdNumber_yAxis.setMaximumSize(QtCore.QSize(120, 30))
-        font = QtGui.QFont()
-        font.setBold(True)
-        font.setWeight(75)
-        self.lcdNumber_yAxis.setFont(font)
-        self.lcdNumber_yAxis.setFrameShadow(QtWidgets.QFrame.Plain)
-        self.lcdNumber_yAxis.setDigitCount(6)
-        self.lcdNumber_yAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
-        self.lcdNumber_yAxis.setObjectName("lcdNumber_yAxis")
-        self.gridLayout_2.addWidget(self.lcdNumber_yAxis, 8, 2, 1, 1)
-        self.pushButton_zHome = QtWidgets.QPushButton(self.frame_stepperControl)
-        self.pushButton_zHome.setMinimumSize(QtCore.QSize(70, 30))
-        self.pushButton_zHome.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.pushButton_zHome.setFont(font)
-        self.pushButton_zHome.setObjectName("pushButton_zHome")
-        self.gridLayout_2.addWidget(self.pushButton_zHome, 10, 2, 1, 1)
-        self.doubleSpinBox_yMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_yMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_yMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_yMin.setFont(font)
-        self.doubleSpinBox_yMin.setDecimals(0)
-        self.doubleSpinBox_yMin.setMinimum(0.0)
-        self.doubleSpinBox_yMin.setMaximum(1599.0)
-        self.doubleSpinBox_yMin.setSingleStep(5.0)
-        self.doubleSpinBox_yMin.setProperty("value", 0.0)
-        self.doubleSpinBox_yMin.setObjectName("doubleSpinBox_yMin")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_yMin, 8, 4, 1, 1)
-        self.doubleSpinBox_xStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_xStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_xStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_xStep.setFont(font)
-        self.doubleSpinBox_xStep.setDecimals(0)
-        self.doubleSpinBox_xStep.setMinimum(0.0)
-        self.doubleSpinBox_xStep.setMaximum(450.0)
-        self.doubleSpinBox_xStep.setSingleStep(5.0)
-        self.doubleSpinBox_xStep.setProperty("value", 50.0)
-        self.doubleSpinBox_xStep.setObjectName("doubleSpinBox_xStep")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_xStep, 5, 5, 1, 1)
-        self.lcdNumber_zAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
-        self.lcdNumber_zAxis.setEnabled(True)
-        self.lcdNumber_zAxis.setMinimumSize(QtCore.QSize(70, 0))
-        self.lcdNumber_zAxis.setMaximumSize(QtCore.QSize(120, 30))
-        font = QtGui.QFont()
-        font.setBold(True)
-        font.setWeight(75)
-        self.lcdNumber_zAxis.setFont(font)
-        self.lcdNumber_zAxis.setFrameShadow(QtWidgets.QFrame.Plain)
-        self.lcdNumber_zAxis.setSmallDecimalPoint(True)
-        self.lcdNumber_zAxis.setDigitCount(6)
-        self.lcdNumber_zAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
-        self.lcdNumber_zAxis.setProperty("intValue", 0)
-        self.lcdNumber_zAxis.setObjectName("lcdNumber_zAxis")
-        self.gridLayout_2.addWidget(self.lcdNumber_zAxis, 11, 2, 1, 1)
-        self.doubleSpinBox_yStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_yStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_yStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_yStep.setFont(font)
-        self.doubleSpinBox_yStep.setDecimals(0)
-        self.doubleSpinBox_yStep.setMinimum(0.0)
-        self.doubleSpinBox_yStep.setMaximum(1600.0)
-        self.doubleSpinBox_yStep.setSingleStep(5.0)
-        self.doubleSpinBox_yStep.setProperty("value", 80.0)
-        self.doubleSpinBox_yStep.setObjectName("doubleSpinBox_yStep")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_yStep, 8, 5, 1, 1)
-        self.doubleSpinBox_yMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_yMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_yMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_yMax.setFont(font)
-        self.doubleSpinBox_yMax.setDecimals(0)
-        self.doubleSpinBox_yMax.setMinimum(0.0)
-        self.doubleSpinBox_yMax.setMaximum(1600.0)
-        self.doubleSpinBox_yMax.setSingleStep(5.0)
-        self.doubleSpinBox_yMax.setProperty("value", 1600.0)
-        self.doubleSpinBox_yMax.setObjectName("doubleSpinBox_yMax")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_yMax, 8, 6, 1, 1)
-        self.doubleSpinBox_zMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_zMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_zMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_zMax.setFont(font)
-        self.doubleSpinBox_zMax.setDecimals(0)
-        self.doubleSpinBox_zMax.setMinimum(-35000.0)
-        self.doubleSpinBox_zMax.setMaximum(0.0)
-        self.doubleSpinBox_zMax.setSingleStep(1000.0)
-        self.doubleSpinBox_zMax.setProperty("value", -8000.0)
-        self.doubleSpinBox_zMax.setObjectName("doubleSpinBox_zMax")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_zMax, 11, 6, 1, 1)
-        self.lcdNumber_xAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
-        self.lcdNumber_xAxis.setMinimumSize(QtCore.QSize(70, 0))
-        self.lcdNumber_xAxis.setMaximumSize(QtCore.QSize(120, 30))
-        font = QtGui.QFont()
-        font.setBold(True)
-        font.setWeight(75)
-        self.lcdNumber_xAxis.setFont(font)
-        self.lcdNumber_xAxis.setFrameShadow(QtWidgets.QFrame.Plain)
-        self.lcdNumber_xAxis.setDigitCount(6)
-        self.lcdNumber_xAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
-        self.lcdNumber_xAxis.setObjectName("lcdNumber_xAxis")
-        self.gridLayout_2.addWidget(self.lcdNumber_xAxis, 5, 2, 1, 1)
-        self.label_xAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_xAxisSliderMax.setFont(font)
-        self.label_xAxisSliderMax.setAutoFillBackground(False)
-        self.label_xAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.label_xAxisSliderMax.setObjectName("label_xAxisSliderMax")
-        self.gridLayout_2.addWidget(self.label_xAxisSliderMax, 4, 1, 1, 1)
-        self.horizontalSlider_yAxis = QtWidgets.QSlider(self.frame_stepperControl)
-        self.horizontalSlider_yAxis.setMinimumSize(QtCore.QSize(500, 0))
-        self.horizontalSlider_yAxis.setMinimum(-1600)
-        self.horizontalSlider_yAxis.setMaximum(1600)
-        self.horizontalSlider_yAxis.setOrientation(QtCore.Qt.Horizontal)
-        self.horizontalSlider_yAxis.setObjectName("horizontalSlider_yAxis")
-        self.gridLayout_2.addWidget(self.horizontalSlider_yAxis, 8, 0, 1, 2)
-        self.doubleSpinBox_zStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_zStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_zStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_zStep.setFont(font)
-        self.doubleSpinBox_zStep.setDecimals(0)
-        self.doubleSpinBox_zStep.setMinimum(50.0)
-        self.doubleSpinBox_zStep.setMaximum(10000.0)
-        self.doubleSpinBox_zStep.setSingleStep(50.0)
-        self.doubleSpinBox_zStep.setProperty("value", 500.0)
-        self.doubleSpinBox_zStep.setObjectName("doubleSpinBox_zStep")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_zStep, 11, 5, 1, 1)
-        self.label_zAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_zAxisSliderMin.setFont(font)
-        self.label_zAxisSliderMin.setObjectName("label_zAxisSliderMin")
-        self.gridLayout_2.addWidget(self.label_zAxisSliderMin, 10, 0, 1, 1)
-        self.label_xAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_xAxisSliderMin.setFont(font)
-        self.label_xAxisSliderMin.setObjectName("label_xAxisSliderMin")
-        self.gridLayout_2.addWidget(self.label_xAxisSliderMin, 4, 0, 1, 1)
-        self.label_yAxis = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_yAxis.setFont(font)
-        self.label_yAxis.setObjectName("label_yAxis")
-        self.gridLayout_2.addWidget(self.label_yAxis, 6, 0, 1, 1)
-        self.pushButton_yReset = QtWidgets.QPushButton(self.frame_stepperControl)
-        self.pushButton_yReset.setMinimumSize(QtCore.QSize(70, 30))
-        self.pushButton_yReset.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.pushButton_yReset.setFont(font)
-        self.pushButton_yReset.setObjectName("pushButton_yReset")
-        self.gridLayout_2.addWidget(self.pushButton_yReset, 7, 2, 1, 1)
-        self.label_stepperControl = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(11)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_stepperControl.setFont(font)
-        self.label_stepperControl.setObjectName("label_stepperControl")
-        self.gridLayout_2.addWidget(self.label_stepperControl, 1, 0, 1, 1)
-        self.doubleSpinBox_xMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
-        self.doubleSpinBox_xMax.setMinimumSize(QtCore.QSize(50, 0))
-        self.doubleSpinBox_xMax.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.doubleSpinBox_xMax.setFont(font)
-        self.doubleSpinBox_xMax.setDecimals(0)
-        self.doubleSpinBox_xMax.setMinimum(0.0)
-        self.doubleSpinBox_xMax.setMaximum(450.0)
-        self.doubleSpinBox_xMax.setSingleStep(5.0)
-        self.doubleSpinBox_xMax.setProperty("value", 450.0)
-        self.doubleSpinBox_xMax.setObjectName("doubleSpinBox_xMax")
-        self.gridLayout_2.addWidget(self.doubleSpinBox_xMax, 5, 6, 1, 1)
-        self.label_setRangeYStep = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeYStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeYStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeYStep.setFont(font)
-        self.label_setRangeYStep.setObjectName("label_setRangeYStep")
-        self.gridLayout_2.addWidget(self.label_setRangeYStep, 7, 5, 1, 1)
-        self.label_setRangeZStep = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeZStep.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeZStep.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeZStep.setFont(font)
-        self.label_setRangeZStep.setObjectName("label_setRangeZStep")
-        self.gridLayout_2.addWidget(self.label_setRangeZStep, 10, 5, 1, 1)
-        self.label_setRangeXMin = QtWidgets.QLabel(self.frame_stepperControl)
-        self.label_setRangeXMin.setMinimumSize(QtCore.QSize(50, 0))
-        self.label_setRangeXMin.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_setRangeXMin.setFont(font)
-        self.label_setRangeXMin.setObjectName("label_setRangeXMin")
-        self.gridLayout_2.addWidget(self.label_setRangeXMin, 4, 4, 1, 1)
-        self.pushButton_xHome = QtWidgets.QPushButton(self.frame_stepperControl)
-        self.pushButton_xHome.setMinimumSize(QtCore.QSize(70, 30))
-        self.pushButton_xHome.setMaximumSize(QtCore.QSize(120, 16777215))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.pushButton_xHome.setFont(font)
-        self.pushButton_xHome.setObjectName("pushButton_xHome")
-        self.gridLayout_2.addWidget(self.pushButton_xHome, 4, 2, 1, 1)
-        self.label_zAxis = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_zAxis.setFont(font)
-        self.label_zAxis.setObjectName("label_zAxis")
-        self.gridLayout_2.addWidget(self.label_zAxis, 9, 0, 1, 1)
-        self.label_zAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_zAxisSliderMax.setFont(font)
-        self.label_zAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.label_zAxisSliderMax.setObjectName("label_zAxisSliderMax")
-        self.gridLayout_2.addWidget(self.label_zAxisSliderMax, 10, 1, 1, 1)
-        self.label_yAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.label_yAxisSliderMin.setFont(font)
-        self.label_yAxisSliderMin.setObjectName("label_yAxisSliderMin")
-        self.gridLayout_2.addWidget(self.label_yAxisSliderMin, 7, 0, 1, 1)
-        self.label_setRange = QtWidgets.QLabel(self.frame_stepperControl)
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_setRange.setFont(font)
-        self.label_setRange.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_setRange.setObjectName("label_setRange")
-        self.gridLayout_2.addWidget(self.label_setRange, 3, 5, 1, 1)
-        self.pushButton_stepperDeEnergise = QtWidgets.QPushButton(self.frame_stepperControl)
-        self.pushButton_stepperDeEnergise.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(True)
-        font.setWeight(75)
-        self.pushButton_stepperDeEnergise.setFont(font)
-        self.pushButton_stepperDeEnergise.setStyleSheet("background-color: rgb(255, 76, 79)")
-        self.pushButton_stepperDeEnergise.setAutoDefault(False)
-        self.pushButton_stepperDeEnergise.setObjectName("pushButton_stepperDeEnergise")
-        self.gridLayout_2.addWidget(self.pushButton_stepperDeEnergise, 1, 5, 1, 1)
-        self.pushButton_Energise = QtWidgets.QPushButton(self.frame_stepperControl)
-        self.pushButton_Energise.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(True)
-        font.setWeight(75)
-        self.pushButton_Energise.setFont(font)
-        self.pushButton_Energise.setAutoFillBackground(False)
-        self.pushButton_Energise.setStyleSheet("background-color: rgb(85, 255, 20)")
-        self.pushButton_Energise.setObjectName("pushButton_Energise")
-        self.gridLayout_2.addWidget(self.pushButton_Energise, 1, 4, 1, 1)
-        self.gridLayout.addWidget(self.frame_stepperControl, 2, 0, 1, 2, QtCore.Qt.AlignTop)
-        self.frame = QtWidgets.QFrame(self.centralwidget)
-        self.frame.setMinimumSize(QtCore.QSize(720, 0))
-        self.frame.setFrameShape(QtWidgets.QFrame.Box)
-        self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame.setObjectName("frame")
-        self.gridLayout_6 = QtWidgets.QGridLayout(self.frame)
-        self.gridLayout_6.setObjectName("gridLayout_6")
-        self.label_liveView = QtWidgets.QLabel(self.frame)
-        self.label_liveView.setEnabled(True)
-        self.label_liveView.setMinimumSize(QtCore.QSize(690, 460))
-        font = QtGui.QFont()
-        font.setPointSize(11)
-        font.setBold(True)
-        font.setItalic(False)
-        font.setWeight(75)
-        self.label_liveView.setFont(font)
-        self.label_liveView.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_liveView.setObjectName("label_liveView")
-        self.gridLayout_6.addWidget(self.label_liveView, 0, 0, 1, 1)
-        self.gridLayout.addWidget(self.frame, 1, 0, 1, 1)
-        self.frame_scannerSetup = QtWidgets.QFrame(self.centralwidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.frame_scannerSetup.sizePolicy().hasHeightForWidth())
-        self.frame_scannerSetup.setSizePolicy(sizePolicy)
-        self.frame_scannerSetup.setMinimumSize(QtCore.QSize(370, 0))
-        self.frame_scannerSetup.setAutoFillBackground(False)
-        self.frame_scannerSetup.setFrameShape(QtWidgets.QFrame.Box)
-        self.frame_scannerSetup.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_scannerSetup.setObjectName("frame_scannerSetup")
-        self.gridLayout_4 = QtWidgets.QGridLayout(self.frame_scannerSetup)
-        self.gridLayout_4.setObjectName("gridLayout_4")
-        self.label_thresholdMasking = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_thresholdMasking.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_thresholdMasking.setFont(font)
-        self.label_thresholdMasking.setObjectName("label_thresholdMasking")
-        self.gridLayout_4.addWidget(self.label_thresholdMasking, 11, 0, 1, 1)
-        self.label_outputFolder = QtWidgets.QLabel(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_outputFolder.setFont(font)
-        self.label_outputFolder.setObjectName("label_outputFolder")
-        self.gridLayout_4.addWidget(self.label_outputFolder, 2, 0, 1, 1)
-        self.lineEdit_projectName = QtWidgets.QLineEdit(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.lineEdit_projectName.setFont(font)
-        self.lineEdit_projectName.setObjectName("lineEdit_projectName")
-        self.gridLayout_4.addWidget(self.lineEdit_projectName, 4, 1, 1, 2)
-        self.label_scannerSetup = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_scannerSetup.setMaximumSize(QtCore.QSize(16777215, 15))
-        font = QtGui.QFont()
-        font.setPointSize(11)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_scannerSetup.setFont(font)
-        self.label_scannerSetup.setLayoutDirection(QtCore.Qt.LeftToRight)
-        self.label_scannerSetup.setObjectName("label_scannerSetup")
-        self.gridLayout_4.addWidget(self.label_scannerSetup, 1, 0, 1, 1)
-        self.label_maskImages = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_maskImages.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_maskImages.setFont(font)
-        self.label_maskImages.setObjectName("label_maskImages")
-        self.gridLayout_4.addWidget(self.label_maskImages, 9, 0, 1, 1)
-        self.pushButton_browseOutput = QtWidgets.QPushButton(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.pushButton_browseOutput.setFont(font)
-        self.pushButton_browseOutput.setObjectName("pushButton_browseOutput")
-        self.gridLayout_4.addWidget(self.pushButton_browseOutput, 2, 1, 1, 2)
-        self.spinBox_thresholdMax = QtWidgets.QSpinBox(self.frame_scannerSetup)
-        self.spinBox_thresholdMax.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.spinBox_thresholdMax.setFont(font)
-        self.spinBox_thresholdMax.setMaximum(254)
-        self.spinBox_thresholdMax.setProperty("value", 215)
-        self.spinBox_thresholdMax.setObjectName("spinBox_thresholdMax")
-        self.gridLayout_4.addWidget(self.spinBox_thresholdMax, 11, 2, 1, 1)
-        self.label_projectName = QtWidgets.QLabel(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_projectName.setFont(font)
-        self.label_projectName.setObjectName("label_projectName")
-        self.gridLayout_4.addWidget(self.label_projectName, 4, 0, 1, 1)
-        self.label_stackImages = QtWidgets.QLabel(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_stackImages.setFont(font)
-        self.label_stackImages.setObjectName("label_stackImages")
-        self.gridLayout_4.addWidget(self.label_stackImages, 6, 0, 1, 1)
-        self.label_thresholdMaskingMin = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_thresholdMaskingMin.setEnabled(False)
-        self.label_thresholdMaskingMin.setMaximumSize(QtCore.QSize(16777215, 20))
-        self.label_thresholdMaskingMin.setObjectName("label_thresholdMaskingMin")
-        self.gridLayout_4.addWidget(self.label_thresholdMaskingMin, 10, 1, 1, 1)
-        self.pushButton_browsePresets = QtWidgets.QPushButton(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.pushButton_browsePresets.setFont(font)
-        self.pushButton_browsePresets.setObjectName("pushButton_browsePresets")
-        self.gridLayout_4.addWidget(self.pushButton_browsePresets, 5, 1, 1, 2)
-        self.comboBox_stackingMethod = QtWidgets.QComboBox(self.frame_scannerSetup)
-        self.comboBox_stackingMethod.setEnabled(False)
-        self.comboBox_stackingMethod.setObjectName("comboBox_stackingMethod")
-        self.comboBox_stackingMethod.addItem("")
-        self.comboBox_stackingMethod.addItem("")
-        self.comboBox_stackingMethod.addItem("")
-        self.gridLayout_4.addWidget(self.comboBox_stackingMethod, 7, 1, 1, 2)
-        self.label_stackingMethod = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_stackingMethod.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_stackingMethod.setFont(font)
-        self.label_stackingMethod.setObjectName("label_stackingMethod")
-        self.gridLayout_4.addWidget(self.label_stackingMethod, 7, 0, 1, 1)
-        self.spinBox_thresholdMin = QtWidgets.QSpinBox(self.frame_scannerSetup)
-        self.spinBox_thresholdMin.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.spinBox_thresholdMin.setFont(font)
-        self.spinBox_thresholdMin.setMaximum(254)
-        self.spinBox_thresholdMin.setProperty("value", 175)
-        self.spinBox_thresholdMin.setObjectName("spinBox_thresholdMin")
-        self.gridLayout_4.addWidget(self.spinBox_thresholdMin, 11, 1, 1, 1)
-        self.label_thresholdMaskingMax = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_thresholdMaskingMax.setEnabled(False)
-        self.label_thresholdMaskingMax.setMaximumSize(QtCore.QSize(16777215, 20))
-        self.label_thresholdMaskingMax.setObjectName("label_thresholdMaskingMax")
-        self.gridLayout_4.addWidget(self.label_thresholdMaskingMax, 10, 2, 1, 1)
-        self.label_thresholdFocus = QtWidgets.QLabel(self.frame_scannerSetup)
-        self.label_thresholdFocus.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_thresholdFocus.setFont(font)
-        self.label_thresholdFocus.setObjectName("label_thresholdFocus")
-        self.gridLayout_4.addWidget(self.label_thresholdFocus, 8, 0, 1, 1)
-        self.label_loadPresets = QtWidgets.QLabel(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_loadPresets.setFont(font)
-        self.label_loadPresets.setObjectName("label_loadPresets")
-        self.gridLayout_4.addWidget(self.label_loadPresets, 5, 0, 1, 1)
-        self.checkBox_stackImages = QtWidgets.QCheckBox(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.checkBox_stackImages.setFont(font)
-        self.checkBox_stackImages.setText("")
-        self.checkBox_stackImages.setObjectName("checkBox_stackImages")
-        self.gridLayout_4.addWidget(self.checkBox_stackImages, 6, 1, 1, 1)
-        self.spinBox_thresholdFocus = QtWidgets.QSpinBox(self.frame_scannerSetup)
-        self.spinBox_thresholdFocus.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.spinBox_thresholdFocus.setFont(font)
-        self.spinBox_thresholdFocus.setMaximum(500)
-        self.spinBox_thresholdFocus.setProperty("value", 10)
-        self.spinBox_thresholdFocus.setObjectName("spinBox_thresholdFocus")
-        self.gridLayout_4.addWidget(self.spinBox_thresholdFocus, 8, 1, 1, 1)
-        self.checkBox_maskImages = QtWidgets.QCheckBox(self.frame_scannerSetup)
-        self.checkBox_maskImages.setEnabled(False)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.checkBox_maskImages.setFont(font)
-        self.checkBox_maskImages.setText("")
-        self.checkBox_maskImages.setObjectName("checkBox_maskImages")
-        self.gridLayout_4.addWidget(self.checkBox_maskImages, 9, 1, 1, 1)
-        self.pushButton_saveConfig = QtWidgets.QPushButton(self.frame_scannerSetup)
-        self.pushButton_saveConfig.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.pushButton_saveConfig.setFont(font)
-        self.pushButton_saveConfig.setObjectName("pushButton_saveConfig")
-        self.gridLayout_4.addWidget(self.pushButton_saveConfig, 12, 2, 1, 1)
-        self.pushButton_startScan = QtWidgets.QPushButton(self.frame_scannerSetup)
-        self.pushButton_startScan.setMinimumSize(QtCore.QSize(0, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.pushButton_startScan.setFont(font)
-        self.pushButton_startScan.setObjectName("pushButton_startScan")
-        self.gridLayout_4.addWidget(self.pushButton_startScan, 12, 0, 1, 2)
-        self.lineEdit_outputLocation = QtWidgets.QLineEdit(self.frame_scannerSetup)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setItalic(True)
-        self.lineEdit_outputLocation.setFont(font)
-        self.lineEdit_outputLocation.setObjectName("lineEdit_outputLocation")
-        self.gridLayout_4.addWidget(self.lineEdit_outputLocation, 3, 1, 1, 2)
-        self.gridLayout.addWidget(self.frame_scannerSetup, 1, 2, 1, 1)
-        MainWindow.setCentralWidget(self.centralwidget)
-        self.statusbar = QtWidgets.QStatusBar(MainWindow)
-        self.statusbar.setObjectName("statusbar")
-        MainWindow.setStatusBar(self.statusbar)
-        self.menubar = QtWidgets.QMenuBar(MainWindow)
-        self.menubar.setGeometry(QtCore.QRect(0, 0, 1600, 18))
-        self.menubar.setObjectName("menubar")
-        self.menuScanner = QtWidgets.QMenu(self.menubar)
-        self.menuScanner.setObjectName("menuScanner")
-        MainWindow.setMenuBar(self.menubar)
-        self.menubar.addAction(self.menuScanner.menuAction())
-
-        self.retranslateUi(MainWindow)
-        self.stacked_camera_settings.setCurrentIndex(0)
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
-
-    def retranslateUi(self, MainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        self.label_info.setText(_translate("MainWindow", "Info"))
-        self.label_progressTotal.setText(_translate("MainWindow", "Progress"))
-        self.label_log.setText(_translate("MainWindow", "Log"))
-        self.pushButton_captureImage.setText(_translate("MainWindow", "Capture Image"))
-        self.label_cameraSettings.setText(_translate("MainWindow", "Camera Settings"))
-        self.pushButton_startLiveView.setText(_translate("MainWindow", "Start Live View"))
-        self.label_selectCamera.setText(_translate("MainWindow", "  Select Camera"))
-        self.label_exposureAuto.setText(_translate("MainWindow", "Exposure Auto"))
-        self.label_exposureTime.setText(_translate("MainWindow", "Exposure time [us]"))
-        self.label_gainAuto.setText(_translate("MainWindow", "Gain Auto"))
-        self.label_gainLevel.setText(_translate("MainWindow", "Gain Level [dB]"))
-        self.label_gamma.setText(_translate("MainWindow", "Gamma"))
-        self.label_balanceRatioRed.setText(_translate("MainWindow", "Balance Ratio (red)"))
-        self.label_balanceRatioBlue.setText(_translate("MainWindow", "Balance Ratio (blue)"))
-        self.label_highlightExposure.setText(_translate("MainWindow", "Highlight Exposure"))
-        self.label_shutterSpeed.setText(_translate("MainWindow", "Shutter speed [s]"))
-        self.label_aperture.setText(_translate("MainWindow", "Aperture (F Stop)"))
-        self.label_iso.setText(_translate("MainWindow", "ISO"))
-        self.label_whiteBalance.setText(_translate("MainWindow", "White Balance"))
-        self.label_compression.setText(_translate("MainWindow", "Compression"))
-        self.label_yAxisSliderMax.setText(_translate("MainWindow", "turn right"))
-        self.label_setRangeXMax.setText(_translate("MainWindow", "Max [X Axis]"))
-        self.label_setRangeZMax.setText(_translate("MainWindow", "Max [Z Axis]"))
-        self.label_setRangeYMin.setText(_translate("MainWindow", "Min [Y Axis]"))
-        self.label_setRangeZMin.setText(_translate("MainWindow", "Min [Z Axis]"))
-        self.label_setRangeYMax.setText(_translate("MainWindow", "Max [Y Axis]"))
-        self.label_xAxis.setText(_translate("MainWindow", "Pitch [X Axis]"))
-        self.label_setRangeXStep.setText(_translate("MainWindow", "Step [X Axis]"))
-        self.pushButton_zHome.setText(_translate("MainWindow", "Home Focus"))
-        self.label_xAxisSliderMax.setText(_translate("MainWindow", "away from camera"))
-        self.label_zAxisSliderMin.setText(_translate("MainWindow", "back position"))
-        self.label_xAxisSliderMin.setText(_translate("MainWindow", "towards camera"))
-        self.label_yAxis.setText(_translate("MainWindow", "Rotation [Y Axis]"))
-        self.pushButton_yReset.setText(_translate("MainWindow", "Reset Y axis"))
-        self.label_stepperControl.setText(_translate("MainWindow", "Stepper Control"))
-        self.label_setRangeYStep.setText(_translate("MainWindow", "Step [Y Axis]"))
-        self.label_setRangeZStep.setText(_translate("MainWindow", "Step [Z Axis]"))
-        self.label_setRangeXMin.setText(_translate("MainWindow", "Min [X Axis]"))
-        self.pushButton_xHome.setText(_translate("MainWindow", "Home X Axis"))
-        self.label_zAxis.setText(_translate("MainWindow", "Focus [Z Axis]"))
-        self.label_zAxisSliderMax.setText(_translate("MainWindow", "front position"))
-        self.label_yAxisSliderMin.setText(_translate("MainWindow", "turn left"))
-        self.label_setRange.setText(_translate("MainWindow", "Set Scanner Range"))
-        self.pushButton_stepperDeEnergise.setText(_translate("MainWindow", "De-Energise"))
-        self.pushButton_Energise.setText(_translate("MainWindow", "Resume"))
-        self.label_liveView.setText(_translate("MainWindow", "Live view disabled."))
-        self.label_thresholdMasking.setText(_translate("MainWindow", "Threshold (masking)"))
-        self.label_outputFolder.setText(_translate("MainWindow", "Output folder"))
-        self.lineEdit_projectName.setText(_translate("MainWindow", "test_project"))
-        self.label_scannerSetup.setText(_translate("MainWindow", "Scanner Setup"))
-        self.label_maskImages.setText(_translate("MainWindow", "Mask images"))
-        self.pushButton_browseOutput.setText(_translate("MainWindow", "browse"))
-        self.label_projectName.setText(_translate("MainWindow", "Project name"))
-        self.label_stackImages.setText(_translate("MainWindow", "Stack images"))
-        self.label_thresholdMaskingMin.setText(_translate("MainWindow", "min"))
-        self.pushButton_browsePresets.setText(_translate("MainWindow", "browse"))
-        self.comboBox_stackingMethod.setCurrentText(_translate("MainWindow", "default"))
-        self.comboBox_stackingMethod.setItemText(0, _translate("MainWindow", "default"))
-        self.comboBox_stackingMethod.setItemText(1, _translate("MainWindow", "gray projector = 1-star"))
-        self.comboBox_stackingMethod.setItemText(2, _translate("MainWindow", "save masks"))
-        self.label_stackingMethod.setText(_translate("MainWindow", "Stacking method"))
-        self.label_thresholdMaskingMax.setText(_translate("MainWindow", "max"))
-        self.label_thresholdFocus.setText(_translate("MainWindow", "Threshold (focus)"))
-        self.label_loadPresets.setText(_translate("MainWindow", "Load presets"))
-        self.pushButton_saveConfig.setText(_translate("MainWindow", "Save config"))
-        self.pushButton_startScan.setText(_translate("MainWindow", "Start Scan"))
-        self.lineEdit_outputLocation.setText(_translate("MainWindow", "current_folder"))
-        self.menuScanner.setTitle(_translate("MainWindow", "Scanner"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MainWindow = QtWidgets.QMainWindow()
-    ui = Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
