+++ GUI/scAnt_GUI_mw/_GUI_DSLR.py
@@ -1,0 +1,54 @@
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+def setupGUI_DSLR(self):
+    self.DSLR = QtWidgets.QWidget()
+    self.DSLR.setObjectName("DSLR")
+    self.gridLayout_7 = QtWidgets.QGridLayout(self.DSLR)
+    self.gridLayout_7.setObjectName("gridLayout_7")
+    self.label_shutterSpeed = QtWidgets.QLabel(self.DSLR)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_shutterSpeed.setFont(font)
+    self.label_shutterSpeed.setObjectName("label_shutterSpeed")
+    self.gridLayout_7.addWidget(self.label_shutterSpeed, 0, 0, 1, 1)
+    self.comboBox_shutterSpeed = QtWidgets.QComboBox(self.DSLR)
+    self.comboBox_shutterSpeed.setObjectName("comboBox_shutterSpeed")
+    self.gridLayout_7.addWidget(self.comboBox_shutterSpeed, 0, 1, 1, 1)
+    self.label_aperture = QtWidgets.QLabel(self.DSLR)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_aperture.setFont(font)
+    self.label_aperture.setObjectName("label_aperture")
+    self.gridLayout_7.addWidget(self.label_aperture, 1, 0, 1, 1)
+    self.comboBox_aperture = QtWidgets.QComboBox(self.DSLR)
+    self.comboBox_aperture.setObjectName("comboBox_aperture")
+    self.gridLayout_7.addWidget(self.comboBox_aperture, 1, 1, 1, 1)
+    self.label_iso = QtWidgets.QLabel(self.DSLR)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_iso.setFont(font)
+    self.label_iso.setObjectName("label_iso")
+    self.gridLayout_7.addWidget(self.label_iso, 2, 0, 1, 1)
+    self.comboBox_iso = QtWidgets.QComboBox(self.DSLR)
+    self.comboBox_iso.setObjectName("comboBox_iso")
+    self.gridLayout_7.addWidget(self.comboBox_iso, 2, 1, 1, 1)
+    self.label_whiteBalance = QtWidgets.QLabel(self.DSLR)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_whiteBalance.setFont(font)
+    self.label_whiteBalance.setObjectName("label_whiteBalance")
+    self.gridLayout_7.addWidget(self.label_whiteBalance, 3, 0, 1, 1)
+    self.comboBox_whiteBalance = QtWidgets.QComboBox(self.DSLR)
+    self.comboBox_whiteBalance.setObjectName("comboBox_whiteBalance")
+    self.gridLayout_7.addWidget(self.comboBox_whiteBalance, 3, 1, 1, 1)
+    self.label_compression = QtWidgets.QLabel(self.DSLR)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_compression.setFont(font)
+    self.label_compression.setObjectName("label_compression")
+    self.gridLayout_7.addWidget(self.label_compression, 4, 0, 1, 1)
+    self.comboBox_compression = QtWidgets.QComboBox(self.DSLR)
+    self.comboBox_compression.setObjectName("comboBox_compression")
+    self.gridLayout_7.addWidget(self.comboBox_compression, 4, 1, 1, 1)
+    self.stacked_camera_settings.addWidget(self.DSLR)
+    self.gridLayout_3.addWidget(self.stacked_camera_settings, 3, 0, 1, 2)
