+++ GUI/scAnt_GUI_mw/_GUI_Steppers.py
@@ -1,0 +1,384 @@
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+def setupGUI_Steppers(self):
+    self.frame_stepperControl = QtWidgets.QFrame(self.centralwidget)
+    sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
+    sizePolicy.setHorizontalStretch(0)
+    sizePolicy.setVerticalStretch(0)
+    sizePolicy.setHeightForWidth(self.frame_stepperControl.sizePolicy().hasHeightForWidth())
+    self.frame_stepperControl.setSizePolicy(sizePolicy)
+    self.frame_stepperControl.setMinimumSize(QtCore.QSize(0, 350))
+    self.frame_stepperControl.setFrameShape(QtWidgets.QFrame.Box)
+    self.frame_stepperControl.setFrameShadow(QtWidgets.QFrame.Raised)
+    self.frame_stepperControl.setObjectName("frame_stepperControl")
+    self.gridLayout_2 = QtWidgets.QGridLayout(self.frame_stepperControl)
+    self.gridLayout_2.setHorizontalSpacing(10)
+    self.gridLayout_2.setObjectName("gridLayout_2")
+    self.label_yAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_yAxisSliderMax.setFont(font)
+    self.label_yAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
+    self.label_yAxisSliderMax.setObjectName("label_yAxisSliderMax")
+    self.gridLayout_2.addWidget(self.label_yAxisSliderMax, 7, 1, 1, 1)
+    self.label_setRangeXMax = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeXMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeXMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeXMax.setFont(font)
+    self.label_setRangeXMax.setObjectName("label_setRangeXMax")
+    self.gridLayout_2.addWidget(self.label_setRangeXMax, 4, 6, 1, 1)
+    self.horizontalSlider_xAxis = QtWidgets.QSlider(self.frame_stepperControl)
+    self.horizontalSlider_xAxis.setMinimumSize(QtCore.QSize(500, 0))
+    self.horizontalSlider_xAxis.setMaximum(450)
+    self.horizontalSlider_xAxis.setOrientation(QtCore.Qt.Horizontal)
+    self.horizontalSlider_xAxis.setObjectName("horizontalSlider_xAxis")
+    self.gridLayout_2.addWidget(self.horizontalSlider_xAxis, 5, 0, 1, 2)
+    self.doubleSpinBox_xMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_xMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_xMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_xMin.setFont(font)
+    self.doubleSpinBox_xMin.setDecimals(0)
+    self.doubleSpinBox_xMin.setMinimum(0.0)
+    self.doubleSpinBox_xMin.setMaximum(450.0)
+    self.doubleSpinBox_xMin.setSingleStep(5.0)
+    self.doubleSpinBox_xMin.setProperty("value", 0.0)
+    self.doubleSpinBox_xMin.setObjectName("doubleSpinBox_xMin")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_xMin, 5, 4, 1, 1)
+    self.label_setRangeZMax = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeZMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeZMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeZMax.setFont(font)
+    self.label_setRangeZMax.setObjectName("label_setRangeZMax")
+    self.gridLayout_2.addWidget(self.label_setRangeZMax, 10, 6, 1, 1)
+    self.label_setRangeYMin = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeYMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeYMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeYMin.setFont(font)
+    self.label_setRangeYMin.setObjectName("label_setRangeYMin")
+    self.gridLayout_2.addWidget(self.label_setRangeYMin, 7, 4, 1, 1)
+    self.doubleSpinBox_zMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_zMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_zMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_zMin.setFont(font)
+    self.doubleSpinBox_zMin.setDecimals(0)
+    self.doubleSpinBox_zMin.setMinimum(-45000.0)
+    self.doubleSpinBox_zMin.setMaximum(0.0)
+    self.doubleSpinBox_zMin.setSingleStep(1000.0)
+    self.doubleSpinBox_zMin.setProperty("value", -20000.0)
+    self.doubleSpinBox_zMin.setObjectName("doubleSpinBox_zMin")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_zMin, 11, 4, 1, 1)
+    self.label_setRangeZMin = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeZMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeZMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeZMin.setFont(font)
+    self.label_setRangeZMin.setObjectName("label_setRangeZMin")
+    self.gridLayout_2.addWidget(self.label_setRangeZMin, 10, 4, 1, 1)
+    self.label_setRangeYMax = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeYMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeYMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeYMax.setFont(font)
+    self.label_setRangeYMax.setObjectName("label_setRangeYMax")
+    self.gridLayout_2.addWidget(self.label_setRangeYMax, 7, 6, 1, 1)
+    self.label_xAxis = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(12)
+    self.label_xAxis.setFont(font)
+    self.label_xAxis.setObjectName("label_xAxis")
+    self.gridLayout_2.addWidget(self.label_xAxis, 3, 0, 1, 1)
+    self.label_setRangeXStep = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeXStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeXStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeXStep.setFont(font)
+    self.label_setRangeXStep.setObjectName("label_setRangeXStep")
+    self.gridLayout_2.addWidget(self.label_setRangeXStep, 4, 5, 1, 1)
+    self.horizontalSlider_zAxis = QtWidgets.QSlider(self.frame_stepperControl)
+    self.horizontalSlider_zAxis.setMinimumSize(QtCore.QSize(500, 0))
+    self.horizontalSlider_zAxis.setMinimum(-45000)
+    self.horizontalSlider_zAxis.setMaximum(0)
+    self.horizontalSlider_zAxis.setOrientation(QtCore.Qt.Horizontal)
+    self.horizontalSlider_zAxis.setObjectName("horizontalSlider_zAxis")
+    self.gridLayout_2.addWidget(self.horizontalSlider_zAxis, 11, 0, 1, 2)
+    self.lcdNumber_yAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
+    self.lcdNumber_yAxis.setMinimumSize(QtCore.QSize(70, 0))
+    self.lcdNumber_yAxis.setMaximumSize(QtCore.QSize(120, 30))
+    font = QtGui.QFont()
+    font.setBold(True)
+    font.setWeight(75)
+    self.lcdNumber_yAxis.setFont(font)
+    self.lcdNumber_yAxis.setFrameShadow(QtWidgets.QFrame.Plain)
+    self.lcdNumber_yAxis.setDigitCount(6)
+    self.lcdNumber_yAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
+    self.lcdNumber_yAxis.setObjectName("lcdNumber_yAxis")
+    self.gridLayout_2.addWidget(self.lcdNumber_yAxis, 8, 2, 1, 1)
+    self.pushButton_zHome = QtWidgets.QPushButton(self.frame_stepperControl)
+    self.pushButton_zHome.setMinimumSize(QtCore.QSize(70, 30))
+    self.pushButton_zHome.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.pushButton_zHome.setFont(font)
+    self.pushButton_zHome.setObjectName("pushButton_zHome")
+    self.gridLayout_2.addWidget(self.pushButton_zHome, 10, 2, 1, 1)
+    self.doubleSpinBox_yMin = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_yMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_yMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_yMin.setFont(font)
+    self.doubleSpinBox_yMin.setDecimals(0)
+    self.doubleSpinBox_yMin.setMinimum(0.0)
+    self.doubleSpinBox_yMin.setMaximum(1599.0)
+    self.doubleSpinBox_yMin.setSingleStep(5.0)
+    self.doubleSpinBox_yMin.setProperty("value", 0.0)
+    self.doubleSpinBox_yMin.setObjectName("doubleSpinBox_yMin")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_yMin, 8, 4, 1, 1)
+    self.doubleSpinBox_xStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_xStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_xStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_xStep.setFont(font)
+    self.doubleSpinBox_xStep.setDecimals(0)
+    self.doubleSpinBox_xStep.setMinimum(0.0)
+    self.doubleSpinBox_xStep.setMaximum(450.0)
+    self.doubleSpinBox_xStep.setSingleStep(5.0)
+    self.doubleSpinBox_xStep.setProperty("value", 50.0)
+    self.doubleSpinBox_xStep.setObjectName("doubleSpinBox_xStep")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_xStep, 5, 5, 1, 1)
+    self.lcdNumber_zAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
+    self.lcdNumber_zAxis.setEnabled(True)
+    self.lcdNumber_zAxis.setMinimumSize(QtCore.QSize(70, 0))
+    self.lcdNumber_zAxis.setMaximumSize(QtCore.QSize(120, 30))
+    font = QtGui.QFont()
+    font.setBold(True)
+    font.setWeight(75)
+    self.lcdNumber_zAxis.setFont(font)
+    self.lcdNumber_zAxis.setFrameShadow(QtWidgets.QFrame.Plain)
+    self.lcdNumber_zAxis.setSmallDecimalPoint(True)
+    self.lcdNumber_zAxis.setDigitCount(6)
+    self.lcdNumber_zAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
+    self.lcdNumber_zAxis.setProperty("intValue", 0)
+    self.lcdNumber_zAxis.setObjectName("lcdNumber_zAxis")
+    self.gridLayout_2.addWidget(self.lcdNumber_zAxis, 11, 2, 1, 1)
+    self.doubleSpinBox_yStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_yStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_yStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_yStep.setFont(font)
+    self.doubleSpinBox_yStep.setDecimals(0)
+    self.doubleSpinBox_yStep.setMinimum(0.0)
+    self.doubleSpinBox_yStep.setMaximum(1600.0)
+    self.doubleSpinBox_yStep.setSingleStep(5.0)
+    self.doubleSpinBox_yStep.setProperty("value", 80.0)
+    self.doubleSpinBox_yStep.setObjectName("doubleSpinBox_yStep")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_yStep, 8, 5, 1, 1)
+    self.doubleSpinBox_yMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_yMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_yMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_yMax.setFont(font)
+    self.doubleSpinBox_yMax.setDecimals(0)
+    self.doubleSpinBox_yMax.setMinimum(0.0)
+    self.doubleSpinBox_yMax.setMaximum(1600.0)
+    self.doubleSpinBox_yMax.setSingleStep(5.0)
+    self.doubleSpinBox_yMax.setProperty("value", 1600.0)
+    self.doubleSpinBox_yMax.setObjectName("doubleSpinBox_yMax")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_yMax, 8, 6, 1, 1)
+    self.doubleSpinBox_zMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_zMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_zMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_zMax.setFont(font)
+    self.doubleSpinBox_zMax.setDecimals(0)
+    self.doubleSpinBox_zMax.setMinimum(-35000.0)
+    self.doubleSpinBox_zMax.setMaximum(0.0)
+    self.doubleSpinBox_zMax.setSingleStep(1000.0)
+    self.doubleSpinBox_zMax.setProperty("value", -8000.0)
+    self.doubleSpinBox_zMax.setObjectName("doubleSpinBox_zMax")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_zMax, 11, 6, 1, 1)
+    self.lcdNumber_xAxis = QtWidgets.QLCDNumber(self.frame_stepperControl)
+    self.lcdNumber_xAxis.setMinimumSize(QtCore.QSize(70, 0))
+    self.lcdNumber_xAxis.setMaximumSize(QtCore.QSize(120, 30))
+    font = QtGui.QFont()
+    font.setBold(True)
+    font.setWeight(75)
+    self.lcdNumber_xAxis.setFont(font)
+    self.lcdNumber_xAxis.setFrameShadow(QtWidgets.QFrame.Plain)
+    self.lcdNumber_xAxis.setDigitCount(6)
+    self.lcdNumber_xAxis.setSegmentStyle(QtWidgets.QLCDNumber.Flat)
+    self.lcdNumber_xAxis.setObjectName("lcdNumber_xAxis")
+    self.gridLayout_2.addWidget(self.lcdNumber_xAxis, 5, 2, 1, 1)
+    self.label_xAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_xAxisSliderMax.setFont(font)
+    self.label_xAxisSliderMax.setAutoFillBackground(False)
+    self.label_xAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
+    self.label_xAxisSliderMax.setObjectName("label_xAxisSliderMax")
+    self.gridLayout_2.addWidget(self.label_xAxisSliderMax, 4, 1, 1, 1)
+    self.horizontalSlider_yAxis = QtWidgets.QSlider(self.frame_stepperControl)
+    self.horizontalSlider_yAxis.setMinimumSize(QtCore.QSize(500, 0))
+    self.horizontalSlider_yAxis.setMinimum(-1600)
+    self.horizontalSlider_yAxis.setMaximum(1600)
+    self.horizontalSlider_yAxis.setOrientation(QtCore.Qt.Horizontal)
+    self.horizontalSlider_yAxis.setObjectName("horizontalSlider_yAxis")
+    self.gridLayout_2.addWidget(self.horizontalSlider_yAxis, 8, 0, 1, 2)
+    self.doubleSpinBox_zStep = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_zStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_zStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_zStep.setFont(font)
+    self.doubleSpinBox_zStep.setDecimals(0)
+    self.doubleSpinBox_zStep.setMinimum(50.0)
+    self.doubleSpinBox_zStep.setMaximum(10000.0)
+    self.doubleSpinBox_zStep.setSingleStep(50.0)
+    self.doubleSpinBox_zStep.setProperty("value", 500.0)
+    self.doubleSpinBox_zStep.setObjectName("doubleSpinBox_zStep")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_zStep, 11, 5, 1, 1)
+    self.label_zAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_zAxisSliderMin.setFont(font)
+    self.label_zAxisSliderMin.setObjectName("label_zAxisSliderMin")
+    self.gridLayout_2.addWidget(self.label_zAxisSliderMin, 10, 0, 1, 1)
+    self.label_xAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_xAxisSliderMin.setFont(font)
+    self.label_xAxisSliderMin.setObjectName("label_xAxisSliderMin")
+    self.gridLayout_2.addWidget(self.label_xAxisSliderMin, 4, 0, 1, 1)
+    self.label_yAxis = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(12)
+    self.label_yAxis.setFont(font)
+    self.label_yAxis.setObjectName("label_yAxis")
+    self.gridLayout_2.addWidget(self.label_yAxis, 6, 0, 1, 1)
+    self.pushButton_yReset = QtWidgets.QPushButton(self.frame_stepperControl)
+    self.pushButton_yReset.setMinimumSize(QtCore.QSize(70, 30))
+    self.pushButton_yReset.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.pushButton_yReset.setFont(font)
+    self.pushButton_yReset.setObjectName("pushButton_yReset")
+    self.gridLayout_2.addWidget(self.pushButton_yReset, 7, 2, 1, 1)
+    self.label_stepperControl = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(11)
+    font.setBold(True)
+    font.setWeight(75)
+    self.label_stepperControl.setFont(font)
+    self.label_stepperControl.setObjectName("label_stepperControl")
+    self.gridLayout_2.addWidget(self.label_stepperControl, 1, 0, 1, 1)
+    self.doubleSpinBox_xMax = QtWidgets.QDoubleSpinBox(self.frame_stepperControl)
+    self.doubleSpinBox_xMax.setMinimumSize(QtCore.QSize(50, 0))
+    self.doubleSpinBox_xMax.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_xMax.setFont(font)
+    self.doubleSpinBox_xMax.setDecimals(0)
+    self.doubleSpinBox_xMax.setMinimum(0.0)
+    self.doubleSpinBox_xMax.setMaximum(450.0)
+    self.doubleSpinBox_xMax.setSingleStep(5.0)
+    self.doubleSpinBox_xMax.setProperty("value", 450.0)
+    self.doubleSpinBox_xMax.setObjectName("doubleSpinBox_xMax")
+    self.gridLayout_2.addWidget(self.doubleSpinBox_xMax, 5, 6, 1, 1)
+    self.label_setRangeYStep = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeYStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeYStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeYStep.setFont(font)
+    self.label_setRangeYStep.setObjectName("label_setRangeYStep")
+    self.gridLayout_2.addWidget(self.label_setRangeYStep, 7, 5, 1, 1)
+    self.label_setRangeZStep = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeZStep.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeZStep.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeZStep.setFont(font)
+    self.label_setRangeZStep.setObjectName("label_setRangeZStep")
+    self.gridLayout_2.addWidget(self.label_setRangeZStep, 10, 5, 1, 1)
+    self.label_setRangeXMin = QtWidgets.QLabel(self.frame_stepperControl)
+    self.label_setRangeXMin.setMinimumSize(QtCore.QSize(50, 0))
+    self.label_setRangeXMin.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_setRangeXMin.setFont(font)
+    self.label_setRangeXMin.setObjectName("label_setRangeXMin")
+    self.gridLayout_2.addWidget(self.label_setRangeXMin, 4, 4, 1, 1)
+    self.pushButton_xHome = QtWidgets.QPushButton(self.frame_stepperControl)
+    self.pushButton_xHome.setMinimumSize(QtCore.QSize(70, 30))
+    self.pushButton_xHome.setMaximumSize(QtCore.QSize(120, 16777215))
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.pushButton_xHome.setFont(font)
+    self.pushButton_xHome.setObjectName("pushButton_xHome")
+    self.gridLayout_2.addWidget(self.pushButton_xHome, 4, 2, 1, 1)
+    self.label_zAxis = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(12)
+    self.label_zAxis.setFont(font)
+    self.label_zAxis.setObjectName("label_zAxis")
+    self.gridLayout_2.addWidget(self.label_zAxis, 9, 0, 1, 1)
+    self.label_zAxisSliderMax = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_zAxisSliderMax.setFont(font)
+    self.label_zAxisSliderMax.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
+    self.label_zAxisSliderMax.setObjectName("label_zAxisSliderMax")
+    self.gridLayout_2.addWidget(self.label_zAxisSliderMax, 10, 1, 1, 1)
+    self.label_yAxisSliderMin = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    self.label_yAxisSliderMin.setFont(font)
+    self.label_yAxisSliderMin.setObjectName("label_yAxisSliderMin")
+    self.gridLayout_2.addWidget(self.label_yAxisSliderMin, 7, 0, 1, 1)
+    self.label_setRange = QtWidgets.QLabel(self.frame_stepperControl)
+    font = QtGui.QFont()
+    font.setPointSize(12)
+    self.label_setRange.setFont(font)
+    self.label_setRange.setAlignment(QtCore.Qt.AlignCenter)
+    self.label_setRange.setObjectName("label_setRange")
+    self.gridLayout_2.addWidget(self.label_setRange, 3, 5, 1, 1)
+    self.pushButton_stepperDeEnergise = QtWidgets.QPushButton(self.frame_stepperControl)
+    self.pushButton_stepperDeEnergise.setMinimumSize(QtCore.QSize(0, 30))
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    font.setBold(True)
+    font.setWeight(75)
+    self.pushButton_stepperDeEnergise.setFont(font)
+    self.pushButton_stepperDeEnergise.setStyleSheet("background-color: rgb(255, 76, 79)")
+    self.pushButton_stepperDeEnergise.setAutoDefault(False)
+    self.pushButton_stepperDeEnergise.setObjectName("pushButton_stepperDeEnergise")
+    self.gridLayout_2.addWidget(self.pushButton_stepperDeEnergise, 1, 5, 1, 1)
+    self.pushButton_Energise = QtWidgets.QPushButton(self.frame_stepperControl)
+    self.pushButton_Energise.setMinimumSize(QtCore.QSize(0, 30))
+    font = QtGui.QFont()
+    font.setPointSize(9)
+    font.setBold(True)
+    font.setWeight(75)
+    self.pushButton_Energise.setFont(font)
+    self.pushButton_Energise.setAutoFillBackground(False)
+    self.pushButton_Energise.setStyleSheet("background-color: rgb(85, 255, 20)")
+    self.pushButton_Energise.setObjectName("pushButton_Energise")
+    self.gridLayout_2.addWidget(self.pushButton_Energise, 1, 4, 1, 1)
+    self.gridLayout.addWidget(self.frame_stepperControl, 2, 0, 1, 2, QtCore.Qt.AlignTop)
