+++ GUI/scAnt_GUI_mw/_GUI_WEBCAM.py
@@ -1,0 +1,92 @@
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+def setupGUI_WEBCAM(self):
+    self.WEBCAM = QtWidgets.QWidget()
+    self.WEBCAM.setObjectName("WEBCAM")
+    self.gridLayout_8 = QtWidgets.QGridLayout(self.WEBCAM)
+    self.gridLayout_8.setObjectName("gridLayout_8")
+    self.label_exposureAuto = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_exposureAuto.setFont(font)
+    self.label_exposureAuto.setObjectName("label_exposureAuto")
+    self.gridLayout_8.addWidget(self.label_exposureAuto, 0, 0, 1, 1)
+    self.checkBox_exposureAuto = QtWidgets.QCheckBox(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.checkBox_exposureAuto.setFont(font)
+    self.checkBox_exposureAuto.setText("")
+    self.checkBox_exposureAuto.setChecked(False)
+    self.checkBox_exposureAuto.setObjectName("checkBox_exposureAuto")
+    self.gridLayout_8.addWidget(self.checkBox_exposureAuto, 0, 1, 1, 1)
+    self.label_exposureTime = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_exposureTime.setFont(font)
+    self.label_exposureTime.setObjectName("label_exposureTime")
+    self.gridLayout_8.addWidget(self.label_exposureTime, 1, 0, 1, 1)
+    self.doubleSpinBox_exposureTime = QtWidgets.QDoubleSpinBox(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_exposureTime.setFont(font)
+    self.doubleSpinBox_exposureTime.setDecimals(0)
+    self.doubleSpinBox_exposureTime.setMinimum(100.0)
+    self.doubleSpinBox_exposureTime.setMaximum(1000000.0)
+    self.doubleSpinBox_exposureTime.setSingleStep(1000.0)
+    self.doubleSpinBox_exposureTime.setProperty("value", 90000.0)
+    self.doubleSpinBox_exposureTime.setObjectName("doubleSpinBox_exposureTime")
+    self.gridLayout_8.addWidget(self.doubleSpinBox_exposureTime, 1, 1, 1, 1)
+    self.label_gainAuto = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_gainAuto.setFont(font)
+    self.label_gainAuto.setObjectName("label_gainAuto")
+    self.gridLayout_8.addWidget(self.label_gainAuto, 2, 0, 1, 1)
+    self.checkBox_gainAuto = QtWidgets.QCheckBox(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.checkBox_gainAuto.setFont(font)
+    self.checkBox_gainAuto.setText("")
+    self.checkBox_gainAuto.setObjectName("checkBox_gainAuto")
+    self.gridLayout_8.addWidget(self.checkBox_gainAuto, 2, 1, 1, 1)
+    self.label_gainLevel = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_gainLevel.setFont(font)
+    self.label_gainLevel.setObjectName("label_gainLevel")
+    self.gridLayout_8.addWidget(self.label_gainLevel, 3, 0, 1, 1)
+    self.doubleSpinBox_gainLevel = QtWidgets.QDoubleSpinBox(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_gainLevel.setFont(font)
+    self.doubleSpinBox_gainLevel.setDecimals(2)
+    self.doubleSpinBox_gainLevel.setMinimum(0.0)
+    self.doubleSpinBox_gainLevel.setMaximum(25.0)
+    self.doubleSpinBox_gainLevel.setProperty("value", 1.83)
+    self.doubleSpinBox_gainLevel.setObjectName("doubleSpinBox_gainLevel")
+    self.gridLayout_8.addWidget(self.doubleSpinBox_gainLevel, 3, 1, 1, 1)
+    self.label_gamma = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_gamma.setFont(font)
+    self.label_gamma.setObjectName("label_gamma")
+    self.gridLayout_8.addWidget(self.label_gamma, 4, 0, 1, 1)
+    self.doubleSpinBox_gamma = QtWidgets.QDoubleSpinBox(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.doubleSpinBox_gamma.setFont(font)
+    self.doubleSpinBox_gamma.setDecimals(2)
+    self.doubleSpinBox_gamma.setMinimum(0.1)
+    self.doubleSpinBox_gamma.setMaximum(4.0)
+    self.doubleSpinBox_gamma.setSingleStep(0.1)
+    self.doubleSpinBox_gamma.setProperty("value", 0.8)
+    self.doubleSpinBox_gamma.setObjectName("doubleSpinBox_gamma")
+    self.gridLayout_8.addWidget(self.doubleSpinBox_gamma, 4, 1, 1, 1)
+    self.label_balanceRatioRed = QtWidgets.QLabel(self.WEBCAM)
+    font = QtGui.QFont()
+    font.setPointSize(10)
+    self.label_balanceRatioRed.setFont(font)
+    self.label_balanceRatioRed.setObjectName("label_balanceRatioRed")
+    self.gridLayout_8.addWidget(self.label_balanceRatioRed, 5, 0, 1, 1)
+    self.doubleSpinBox_balanceRatioRed = QtWidgets.QDoubleSpinBox(self.WEBCAM)
+    self.stacked_camera_settings.addWidget(self.WEBCAM)
