+++ GUI/scAnt_GUI_mw/__init__.py
@@ -1,0 +1,437 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file 'I:\3D_Scanner\scAnt\GUI\scAnt_GUI_mw_grid.ui'
+#
+# Created by: PyQt5 UI code generator 5.13.0
+#
+# WARNING! All changes made in this file will be lost!
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+from ._GUI_FLIR import setupGUI_FLIR
+from ._GUI_DSLR import setupGUI_DSLR
+from ._GUI_WEBCAM import setupGUI_WEBCAM
+from ._GUI_Steppers import setupGUI_Steppers
+
+class Ui_MainWindow(object):
+    def setupUi(self, MainWindow):
+        MainWindow.setObjectName("MainWindow")
+        MainWindow.setWindowModality(QtCore.Qt.NonModal)
+        MainWindow.setEnabled(True)
+        MainWindow.resize(1600, 936)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(MainWindow.sizePolicy().hasHeightForWidth())
+        MainWindow.setSizePolicy(sizePolicy)
+        MainWindow.setMinimumSize(QtCore.QSize(1600, 936))
+        MainWindow.setMaximumSize(QtCore.QSize(3840, 2160))
+        MainWindow.setWindowTitle("scAnt")
+        self.centralwidget = QtWidgets.QWidget(MainWindow)
+        self.centralwidget.setMinimumSize(QtCore.QSize(1600, 900))
+        self.centralwidget.setObjectName("centralwidget")
+        self.gridLayout = QtWidgets.QGridLayout(self.centralwidget)
+        self.gridLayout.setObjectName("gridLayout")
+        self.frame_info = QtWidgets.QFrame(self.centralwidget)
+        self.frame_info.setMinimumSize(QtCore.QSize(370, 0))
+        self.frame_info.setFrameShape(QtWidgets.QFrame.Box)
+        self.frame_info.setFrameShadow(QtWidgets.QFrame.Raised)
+        self.frame_info.setObjectName("frame_info")
+        self.verticalLayout = QtWidgets.QVBoxLayout(self.frame_info)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.label_info = QtWidgets.QLabel(self.frame_info)
+        self.label_info.setMaximumSize(QtCore.QSize(16777215, 30))
+        font = QtGui.QFont()
+        font.setPointSize(11)
+        font.setBold(True)
+        font.setWeight(75)
+        self.label_info.setFont(font)
+        self.label_info.setObjectName("label_info")
+        self.verticalLayout.addWidget(self.label_info)
+        self.label_progressTotal = QtWidgets.QLabel(self.frame_info)
+        self.label_progressTotal.setEnabled(False)
+        self.label_progressTotal.setMaximumSize(QtCore.QSize(16777215, 30))
+        font = QtGui.QFont()
+        font.setPointSize(12)
+        self.label_progressTotal.setFont(font)
+        self.label_progressTotal.setObjectName("label_progressTotal")
+        self.verticalLayout.addWidget(self.label_progressTotal)
+        self.progressBar_total = QtWidgets.QProgressBar(self.frame_info)
+        self.progressBar_total.setEnabled(False)
+        self.progressBar_total.setMaximumSize(QtCore.QSize(16777215, 40))
+        self.progressBar_total.setProperty("value", 0)
+        self.progressBar_total.setObjectName("progressBar_total")
+        self.verticalLayout.addWidget(self.progressBar_total)
+        self.label_log = QtWidgets.QLabel(self.frame_info)
+        self.label_log.setMaximumSize(QtCore.QSize(16777215, 30))
+        font = QtGui.QFont()
+        font.setPointSize(12)
+        self.label_log.setFont(font)
+        self.label_log.setObjectName("label_log")
+        self.verticalLayout.addWidget(self.label_log)
+        self.listWidget_log = QtWidgets.QListWidget(self.frame_info)
+        self.listWidget_log.setMinimumSize(QtCore.QSize(0, 200))
+        self.listWidget_log.setMaximumSize(QtCore.QSize(1000, 400))
+        font = QtGui.QFont()
+        font.setPointSize(8)
+        self.listWidget_log.setFont(font)
+        self.listWidget_log.setFlow(QtWidgets.QListView.TopToBottom)
+        self.listWidget_log.setObjectName("listWidget_log")
+        self.verticalLayout.addWidget(self.listWidget_log)
+        self.gridLayout.addWidget(self.frame_info, 2, 2, 1, 1)
+        self.frame_cameraSettings = QtWidgets.QFrame(self.centralwidget)
+        self.frame_cameraSettings.setMinimumSize(QtCore.QSize(370, 0))
+        self.frame_cameraSettings.setFrameShape(QtWidgets.QFrame.Box)
+        self.frame_cameraSettings.setFrameShadow(QtWidgets.QFrame.Raised)
+        self.frame_cameraSettings.setObjectName("frame_cameraSettings")
+        self.gridLayout_3 = QtWidgets.QGridLayout(self.frame_cameraSettings)
+        self.gridLayout_3.setObjectName("gridLayout_3")
+        self.pushButton_captureImage = QtWidgets.QPushButton(self.frame_cameraSettings)
+        self.pushButton_captureImage.setMinimumSize(QtCore.QSize(0, 30))
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        font.setBold(False)
+        font.setWeight(50)
+        self.pushButton_captureImage.setFont(font)
+        self.pushButton_captureImage.setObjectName("pushButton_captureImage")
+        self.gridLayout_3.addWidget(self.pushButton_captureImage, 4, 1, 1, 1)
+        self.comboBox_selectCamera = QtWidgets.QComboBox(self.frame_cameraSettings)
+        self.comboBox_selectCamera.setObjectName("comboBox_selectCamera")
+        self.gridLayout_3.addWidget(self.comboBox_selectCamera, 2, 1, 1, 1)
+        self.label_cameraSettings = QtWidgets.QLabel(self.frame_cameraSettings)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Minimum)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.label_cameraSettings.sizePolicy().hasHeightForWidth())
+        self.label_cameraSettings.setSizePolicy(sizePolicy)
+        self.label_cameraSettings.setMinimumSize(QtCore.QSize(0, 40))
+        self.label_cameraSettings.setMaximumSize(QtCore.QSize(16777215, 50))
+        font = QtGui.QFont()
+        font.setPointSize(11)
+        font.setBold(True)
+        font.setWeight(75)
+        font.setKerning(True)
+        self.label_cameraSettings.setFont(font)
+        self.label_cameraSettings.setObjectName("label_cameraSettings")
+        self.gridLayout_3.addWidget(self.label_cameraSettings, 0, 0, 1, 2)
+        self.pushButton_startLiveView = QtWidgets.QPushButton(self.frame_cameraSettings)
+        self.pushButton_startLiveView.setMinimumSize(QtCore.QSize(0, 30))
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        font.setBold(False)
+        font.setWeight(50)
+        self.pushButton_startLiveView.setFont(font)
+        self.pushButton_startLiveView.setObjectName("pushButton_startLiveView")
+        self.gridLayout_3.addWidget(self.pushButton_startLiveView, 4, 0, 1, 1)
+        self.label_selectCamera = QtWidgets.QLabel(self.frame_cameraSettings)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_selectCamera.setFont(font)
+        self.label_selectCamera.setObjectName("label_selectCamera")
+        self.gridLayout_3.addWidget(self.label_selectCamera, 2, 0, 1, 1)
+        self.stacked_camera_settings = QtWidgets.QStackedWidget(self.frame_cameraSettings)
+        self.stacked_camera_settings.setAccessibleDescription("")
+        self.stacked_camera_settings.setObjectName("stacked_camera_settings")
+
+        setupGUI_FLIR(self)
+        setupGUI_DSLR(self)
+
+        self.gridLayout.addWidget(self.frame_cameraSettings, 1, 1, 1, 1)
+
+        # setupGUI_WEBCAM(self)
+        setupGUI_Steppers(self)
+
+        self.frame = QtWidgets.QFrame(self.centralwidget)
+        self.frame.setMinimumSize(QtCore.QSize(720, 0))
+        self.frame.setFrameShape(QtWidgets.QFrame.Box)
+        self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
+        self.frame.setObjectName("frame")
+        self.gridLayout_6 = QtWidgets.QGridLayout(self.frame)
+        self.gridLayout_6.setObjectName("gridLayout_6")
+        self.label_liveView = QtWidgets.QLabel(self.frame)
+        self.label_liveView.setEnabled(True)
+        self.label_liveView.setMinimumSize(QtCore.QSize(690, 460))
+        font = QtGui.QFont()
+        font.setPointSize(11)
+        font.setBold(True)
+        font.setItalic(False)
+        font.setWeight(75)
+        self.label_liveView.setFont(font)
+        self.label_liveView.setAlignment(QtCore.Qt.AlignCenter)
+        self.label_liveView.setObjectName("label_liveView")
+        self.gridLayout_6.addWidget(self.label_liveView, 0, 0, 1, 1)
+        self.gridLayout.addWidget(self.frame, 1, 0, 1, 1)
+        self.frame_scannerSetup = QtWidgets.QFrame(self.centralwidget)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.frame_scannerSetup.sizePolicy().hasHeightForWidth())
+        self.frame_scannerSetup.setSizePolicy(sizePolicy)
+        self.frame_scannerSetup.setMinimumSize(QtCore.QSize(370, 0))
+        self.frame_scannerSetup.setAutoFillBackground(False)
+        self.frame_scannerSetup.setFrameShape(QtWidgets.QFrame.Box)
+        self.frame_scannerSetup.setFrameShadow(QtWidgets.QFrame.Raised)
+        self.frame_scannerSetup.setObjectName("frame_scannerSetup")
+        self.gridLayout_4 = QtWidgets.QGridLayout(self.frame_scannerSetup)
+        self.gridLayout_4.setObjectName("gridLayout_4")
+        self.label_thresholdMasking = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_thresholdMasking.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_thresholdMasking.setFont(font)
+        self.label_thresholdMasking.setObjectName("label_thresholdMasking")
+        self.gridLayout_4.addWidget(self.label_thresholdMasking, 11, 0, 1, 1)
+        self.label_outputFolder = QtWidgets.QLabel(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_outputFolder.setFont(font)
+        self.label_outputFolder.setObjectName("label_outputFolder")
+        self.gridLayout_4.addWidget(self.label_outputFolder, 2, 0, 1, 1)
+        self.lineEdit_projectName = QtWidgets.QLineEdit(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.lineEdit_projectName.setFont(font)
+        self.lineEdit_projectName.setObjectName("lineEdit_projectName")
+        self.gridLayout_4.addWidget(self.lineEdit_projectName, 4, 1, 1, 2)
+        self.label_scannerSetup = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_scannerSetup.setMaximumSize(QtCore.QSize(16777215, 15))
+        font = QtGui.QFont()
+        font.setPointSize(11)
+        font.setBold(True)
+        font.setWeight(75)
+        self.label_scannerSetup.setFont(font)
+        self.label_scannerSetup.setLayoutDirection(QtCore.Qt.LeftToRight)
+        self.label_scannerSetup.setObjectName("label_scannerSetup")
+        self.gridLayout_4.addWidget(self.label_scannerSetup, 1, 0, 1, 1)
+        self.label_maskImages = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_maskImages.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_maskImages.setFont(font)
+        self.label_maskImages.setObjectName("label_maskImages")
+        self.gridLayout_4.addWidget(self.label_maskImages, 9, 0, 1, 1)
+        self.pushButton_browseOutput = QtWidgets.QPushButton(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.pushButton_browseOutput.setFont(font)
+        self.pushButton_browseOutput.setObjectName("pushButton_browseOutput")
+        self.gridLayout_4.addWidget(self.pushButton_browseOutput, 2, 1, 1, 2)
+        self.spinBox_thresholdMax = QtWidgets.QSpinBox(self.frame_scannerSetup)
+        self.spinBox_thresholdMax.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(9)
+        self.spinBox_thresholdMax.setFont(font)
+        self.spinBox_thresholdMax.setMaximum(254)
+        self.spinBox_thresholdMax.setProperty("value", 215)
+        self.spinBox_thresholdMax.setObjectName("spinBox_thresholdMax")
+        self.gridLayout_4.addWidget(self.spinBox_thresholdMax, 11, 2, 1, 1)
+        self.label_projectName = QtWidgets.QLabel(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_projectName.setFont(font)
+        self.label_projectName.setObjectName("label_projectName")
+        self.gridLayout_4.addWidget(self.label_projectName, 4, 0, 1, 1)
+        self.label_stackImages = QtWidgets.QLabel(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_stackImages.setFont(font)
+        self.label_stackImages.setObjectName("label_stackImages")
+        self.gridLayout_4.addWidget(self.label_stackImages, 6, 0, 1, 1)
+        self.label_thresholdMaskingMin = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_thresholdMaskingMin.setEnabled(False)
+        self.label_thresholdMaskingMin.setMaximumSize(QtCore.QSize(16777215, 20))
+        self.label_thresholdMaskingMin.setObjectName("label_thresholdMaskingMin")
+        self.gridLayout_4.addWidget(self.label_thresholdMaskingMin, 10, 1, 1, 1)
+        self.pushButton_browsePresets = QtWidgets.QPushButton(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.pushButton_browsePresets.setFont(font)
+        self.pushButton_browsePresets.setObjectName("pushButton_browsePresets")
+        self.gridLayout_4.addWidget(self.pushButton_browsePresets, 5, 1, 1, 2)
+        self.comboBox_stackingMethod = QtWidgets.QComboBox(self.frame_scannerSetup)
+        self.comboBox_stackingMethod.setEnabled(False)
+        self.comboBox_stackingMethod.setObjectName("comboBox_stackingMethod")
+        self.comboBox_stackingMethod.addItem("")
+        self.comboBox_stackingMethod.addItem("")
+        self.comboBox_stackingMethod.addItem("")
+        self.gridLayout_4.addWidget(self.comboBox_stackingMethod, 7, 1, 1, 2)
+        self.label_stackingMethod = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_stackingMethod.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_stackingMethod.setFont(font)
+        self.label_stackingMethod.setObjectName("label_stackingMethod")
+        self.gridLayout_4.addWidget(self.label_stackingMethod, 7, 0, 1, 1)
+        self.spinBox_thresholdMin = QtWidgets.QSpinBox(self.frame_scannerSetup)
+        self.spinBox_thresholdMin.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(9)
+        self.spinBox_thresholdMin.setFont(font)
+        self.spinBox_thresholdMin.setMaximum(254)
+        self.spinBox_thresholdMin.setProperty("value", 175)
+        self.spinBox_thresholdMin.setObjectName("spinBox_thresholdMin")
+        self.gridLayout_4.addWidget(self.spinBox_thresholdMin, 11, 1, 1, 1)
+        self.label_thresholdMaskingMax = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_thresholdMaskingMax.setEnabled(False)
+        self.label_thresholdMaskingMax.setMaximumSize(QtCore.QSize(16777215, 20))
+        self.label_thresholdMaskingMax.setObjectName("label_thresholdMaskingMax")
+        self.gridLayout_4.addWidget(self.label_thresholdMaskingMax, 10, 2, 1, 1)
+        self.label_thresholdFocus = QtWidgets.QLabel(self.frame_scannerSetup)
+        self.label_thresholdFocus.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_thresholdFocus.setFont(font)
+        self.label_thresholdFocus.setObjectName("label_thresholdFocus")
+        self.gridLayout_4.addWidget(self.label_thresholdFocus, 8, 0, 1, 1)
+        self.label_loadPresets = QtWidgets.QLabel(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.label_loadPresets.setFont(font)
+        self.label_loadPresets.setObjectName("label_loadPresets")
+        self.gridLayout_4.addWidget(self.label_loadPresets, 5, 0, 1, 1)
+        self.checkBox_stackImages = QtWidgets.QCheckBox(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.checkBox_stackImages.setFont(font)
+        self.checkBox_stackImages.setText("")
+        self.checkBox_stackImages.setObjectName("checkBox_stackImages")
+        self.gridLayout_4.addWidget(self.checkBox_stackImages, 6, 1, 1, 1)
+        self.spinBox_thresholdFocus = QtWidgets.QSpinBox(self.frame_scannerSetup)
+        self.spinBox_thresholdFocus.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(9)
+        self.spinBox_thresholdFocus.setFont(font)
+        self.spinBox_thresholdFocus.setMaximum(500)
+        self.spinBox_thresholdFocus.setProperty("value", 10)
+        self.spinBox_thresholdFocus.setObjectName("spinBox_thresholdFocus")
+        self.gridLayout_4.addWidget(self.spinBox_thresholdFocus, 8, 1, 1, 1)
+        self.checkBox_maskImages = QtWidgets.QCheckBox(self.frame_scannerSetup)
+        self.checkBox_maskImages.setEnabled(False)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.checkBox_maskImages.setFont(font)
+        self.checkBox_maskImages.setText("")
+        self.checkBox_maskImages.setObjectName("checkBox_maskImages")
+        self.gridLayout_4.addWidget(self.checkBox_maskImages, 9, 1, 1, 1)
+        self.pushButton_saveConfig = QtWidgets.QPushButton(self.frame_scannerSetup)
+        self.pushButton_saveConfig.setMinimumSize(QtCore.QSize(0, 30))
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        font.setBold(False)
+        font.setWeight(50)
+        self.pushButton_saveConfig.setFont(font)
+        self.pushButton_saveConfig.setObjectName("pushButton_saveConfig")
+        self.gridLayout_4.addWidget(self.pushButton_saveConfig, 12, 2, 1, 1)
+        self.pushButton_startScan = QtWidgets.QPushButton(self.frame_scannerSetup)
+        self.pushButton_startScan.setMinimumSize(QtCore.QSize(0, 30))
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        font.setBold(True)
+        font.setWeight(75)
+        self.pushButton_startScan.setFont(font)
+        self.pushButton_startScan.setObjectName("pushButton_startScan")
+        self.gridLayout_4.addWidget(self.pushButton_startScan, 12, 0, 1, 2)
+        self.lineEdit_outputLocation = QtWidgets.QLineEdit(self.frame_scannerSetup)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        font.setItalic(True)
+        self.lineEdit_outputLocation.setFont(font)
+        self.lineEdit_outputLocation.setObjectName("lineEdit_outputLocation")
+        self.gridLayout_4.addWidget(self.lineEdit_outputLocation, 3, 1, 1, 2)
+        self.gridLayout.addWidget(self.frame_scannerSetup, 1, 2, 1, 1)
+        MainWindow.setCentralWidget(self.centralwidget)
+        self.statusbar = QtWidgets.QStatusBar(MainWindow)
+        self.statusbar.setObjectName("statusbar")
+        MainWindow.setStatusBar(self.statusbar)
+        self.menubar = QtWidgets.QMenuBar(MainWindow)
+        self.menubar.setGeometry(QtCore.QRect(0, 0, 1600, 18))
+        self.menubar.setObjectName("menubar")
+        self.menuScanner = QtWidgets.QMenu(self.menubar)
+        self.menuScanner.setObjectName("menuScanner")
+        MainWindow.setMenuBar(self.menubar)
+        self.menubar.addAction(self.menuScanner.menuAction())
+
+        self.retranslateUi(MainWindow)
+        self.stacked_camera_settings.setCurrentIndex(0)
+        QtCore.QMetaObject.connectSlotsByName(MainWindow)
+
+    def retranslateUi(self, MainWindow):
+        _translate = QtCore.QCoreApplication.translate
+        self.label_info.setText(_translate("MainWindow", "Info"))
+        self.label_progressTotal.setText(_translate("MainWindow", "Progress"))
+        self.label_log.setText(_translate("MainWindow", "Log"))
+        self.pushButton_captureImage.setText(_translate("MainWindow", "Capture Image"))
+        self.label_cameraSettings.setText(_translate("MainWindow", "Camera Settings"))
+        self.pushButton_startLiveView.setText(_translate("MainWindow", "Start Live View"))
+        self.label_selectCamera.setText(_translate("MainWindow", "  Select Camera"))
+        self.label_exposureAuto.setText(_translate("MainWindow", "Exposure Auto"))
+        self.label_exposureTime.setText(_translate("MainWindow", "Exposure time [us]"))
+        self.label_gainAuto.setText(_translate("MainWindow", "Gain Auto"))
+        self.label_gainLevel.setText(_translate("MainWindow", "Gain Level [dB]"))
+        self.label_gamma.setText(_translate("MainWindow", "Gamma"))
+        self.label_balanceRatioRed.setText(_translate("MainWindow", "Balance Ratio (red)"))
+        self.label_balanceRatioBlue.setText(_translate("MainWindow", "Balance Ratio (blue)"))
+        self.label_highlightExposure.setText(_translate("MainWindow", "Highlight Exposure"))
+        self.label_shutterSpeed.setText(_translate("MainWindow", "Shutter speed [s]"))
+        self.label_aperture.setText(_translate("MainWindow", "Aperture (F Stop)"))
+        self.label_iso.setText(_translate("MainWindow", "ISO"))
+        self.label_whiteBalance.setText(_translate("MainWindow", "White Balance"))
+        self.label_compression.setText(_translate("MainWindow", "Compression"))
+        self.label_yAxisSliderMax.setText(_translate("MainWindow", "turn right"))
+        self.label_setRangeXMax.setText(_translate("MainWindow", "Max [X Axis]"))
+        self.label_setRangeZMax.setText(_translate("MainWindow", "Max [Z Axis]"))
+        self.label_setRangeYMin.setText(_translate("MainWindow", "Min [Y Axis]"))
+        self.label_setRangeZMin.setText(_translate("MainWindow", "Min [Z Axis]"))
+        self.label_setRangeYMax.setText(_translate("MainWindow", "Max [Y Axis]"))
+        self.label_xAxis.setText(_translate("MainWindow", "Pitch [X Axis]"))
+        self.label_setRangeXStep.setText(_translate("MainWindow", "Step [X Axis]"))
+        self.pushButton_zHome.setText(_translate("MainWindow", "Home Focus"))
+        self.label_xAxisSliderMax.setText(_translate("MainWindow", "away from camera"))
+        self.label_zAxisSliderMin.setText(_translate("MainWindow", "back position"))
+        self.label_xAxisSliderMin.setText(_translate("MainWindow", "towards camera"))
+        self.label_yAxis.setText(_translate("MainWindow", "Rotation [Y Axis]"))
+        self.pushButton_yReset.setText(_translate("MainWindow", "Reset Y axis"))
+        self.label_stepperControl.setText(_translate("MainWindow", "Stepper Control"))
+        self.label_setRangeYStep.setText(_translate("MainWindow", "Step [Y Axis]"))
+        self.label_setRangeZStep.setText(_translate("MainWindow", "Step [Z Axis]"))
+        self.label_setRangeXMin.setText(_translate("MainWindow", "Min [X Axis]"))
+        self.pushButton_xHome.setText(_translate("MainWindow", "Home X Axis"))
+        self.label_zAxis.setText(_translate("MainWindow", "Focus [Z Axis]"))
+        self.label_zAxisSliderMax.setText(_translate("MainWindow", "front position"))
+        self.label_yAxisSliderMin.setText(_translate("MainWindow", "turn left"))
+        self.label_setRange.setText(_translate("MainWindow", "Set Scanner Range"))
+        self.pushButton_stepperDeEnergise.setText(_translate("MainWindow", "De-Energise"))
+        self.pushButton_Energise.setText(_translate("MainWindow", "Resume"))
+        self.label_liveView.setText(_translate("MainWindow", "Live view disabled."))
+        self.label_thresholdMasking.setText(_translate("MainWindow", "Threshold (masking)"))
+        self.label_outputFolder.setText(_translate("MainWindow", "Output folder"))
+        self.lineEdit_projectName.setText(_translate("MainWindow", "test_project"))
+        self.label_scannerSetup.setText(_translate("MainWindow", "Scanner Setup"))
+        self.label_maskImages.setText(_translate("MainWindow", "Mask images"))
+        self.pushButton_browseOutput.setText(_translate("MainWindow", "browse"))
+        self.label_projectName.setText(_translate("MainWindow", "Project name"))
+        self.label_stackImages.setText(_translate("MainWindow", "Stack images"))
+        self.label_thresholdMaskingMin.setText(_translate("MainWindow", "min"))
+        self.pushButton_browsePresets.setText(_translate("MainWindow", "browse"))
+        self.comboBox_stackingMethod.setCurrentText(_translate("MainWindow", "default"))
+        self.comboBox_stackingMethod.setItemText(0, _translate("MainWindow", "default"))
+        self.comboBox_stackingMethod.setItemText(1, _translate("MainWindow", "gray projector = 1-star"))
+        self.comboBox_stackingMethod.setItemText(2, _translate("MainWindow", "save masks"))
+        self.label_stackingMethod.setText(_translate("MainWindow", "Stacking method"))
+        self.label_thresholdMaskingMax.setText(_translate("MainWindow", "max"))
+        self.label_thresholdFocus.setText(_translate("MainWindow", "Threshold (focus)"))
+        self.label_loadPresets.setText(_translate("MainWindow", "Load presets"))
+        self.pushButton_saveConfig.setText(_translate("MainWindow", "Save config"))
+        self.pushButton_startScan.setText(_translate("MainWindow", "Start Scan"))
+        self.lineEdit_outputLocation.setText(_translate("MainWindow", "current_folder"))
+        self.menuScanner.setTitle(_translate("MainWindow", "Scanner"))
+
+
+if __name__ == "__main__":
+    import sys
+    app = QtWidgets.QApplication(sys.argv)
+    MainWindow = QtWidgets.QMainWindow()
+    ui = Ui_MainWindow()
+    ui.setupUi(MainWindow)
+    MainWindow.show()
+    sys.exit(app.exec_())
